@@ -1,16 +1,16 @@
 module github.com/golangci/golangci-lint/scripts/gen_github_action_config
 
-go 1.19
+go 1.22.0
 
 require (
-	github.com/shurcooL/githubv4 v0.0.0-20230215024106-420ad0987b9b
-	golang.org/x/oauth2 v0.14.0
+	github.com/shurcooL/githubv4 v0.0.0-20240429030203-be2daab69064
+	github.com/stretchr/testify v1.10.0
+	golang.org/x/oauth2 v0.26.0
 )
 
 require (
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/shurcooL/graphql v0.0.0-20220606043923-3cf50f8a0a29 // indirect
-	golang.org/x/net v0.18.0 // indirect
-	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/shurcooL/graphql v0.0.0-20230722043721-ed46e5a46466 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
