@@ -1,6 +1,7 @@
 package main
 
 import (
+	"cmp"
 	"fmt"
 	"os"
 	"runtime/debug"
@@ -13,33 +14,61 @@ var (
 	goVersion = "unknown"
 
 	// Populated by goreleaser during build
-	version = "master"
+	version = "unknown"
 	commit  = "?"
 	date    = ""
 )
 
 func main() {
-	if buildInfo, available := debug.ReadBuildInfo(); available {
-		goVersion = buildInfo.GoVersion
+	info := createBuildInfo()
 
-		if date == "" {
-			version = buildInfo.Main.Version
-			commit = fmt.Sprintf("(unknown, mod sum: %q)", buildInfo.Main.Sum)
-			date = "(unknown)"
-		}
+	if err := commands.Execute(info); err != nil {
+		_, _ = fmt.Fprintf(os.Stderr, "Failed executing command with error: %v\n", err)
+		os.Exit(exitcodes.Failure)
 	}
+}
 
+func createBuildInfo() commands.BuildInfo {
 	info := commands.BuildInfo{
-		GoVersion: goVersion,
-		Version:   version,
 		Commit:    commit,
+		Version:   version,
+		GoVersion: goVersion,
 		Date:      date,
 	}
 
-	e := commands.NewExecutor(info)
-
-	if err := e.Execute(); err != nil {
-		fmt.Fprintf(os.Stderr, "failed executing command with error %v\n", err)
-		os.Exit(exitcodes.Failure)
+	buildInfo, available := debug.ReadBuildInfo()
+	if !available {
+		return info
 	}
+
+	info.GoVersion = buildInfo.GoVersion
+
+	if date != "" {
+		return info
+	}
+
+	info.Version = buildInfo.Main.Version
+
+	var revision string
+	var modified string
+	for _, setting := range buildInfo.Settings {
+		// The `vcs.xxx` information is only available with `go build`.
+		// This information is not available with `go install` or `go run`.
+		switch setting.Key {
+		case "vcs.time":
+			info.Date = setting.Value
+		case "vcs.revision":
+			revision = setting.Value
+		case "vcs.modified":
+			modified = setting.Value
+		}
+	}
+
+	revision = cmp.Or(revision, "unknown")
+	modified = cmp.Or(modified, "?")
+	info.Date = cmp.Or(info.Date, "(unknown)")
+
+	info.Commit = fmt.Sprintf("(%s, modified: %s, mod sum: %q)", revision, modified, buildInfo.Main.Sum)
+
+	return info
 }
