@@ -12,7 +12,13 @@ import (
 	"github.com/golangci/golangci-lint/pkg/logutils"
 )
 
-func (e *Executor) initCache() {
+type cacheCommand struct {
+	cmd *cobra.Command
+}
+
+func newCacheCommand() *cacheCommand {
+	c := &cacheCommand{}
+
 	cacheCmd := &cobra.Command{
 		Use:   "cache",
 		Short: "Cache control and information",
@@ -21,28 +27,32 @@ func (e *Executor) initCache() {
 			return cmd.Help()
 		},
 	}
-	e.rootCmd.AddCommand(cacheCmd)
 
-	cacheCmd.AddCommand(&cobra.Command{
-		Use:               "clean",
-		Short:             "Clean cache",
-		Args:              cobra.NoArgs,
-		ValidArgsFunction: cobra.NoFileCompletions,
-		RunE:              e.executeCleanCache,
-	})
-	cacheCmd.AddCommand(&cobra.Command{
-		Use:               "status",
-		Short:             "Show cache status",
-		Args:              cobra.NoArgs,
-		ValidArgsFunction: cobra.NoFileCompletions,
-		Run:               e.executeCacheStatus,
-	})
+	cacheCmd.AddCommand(
+		&cobra.Command{
+			Use:               "clean",
+			Short:             "Clean cache",
+			Args:              cobra.NoArgs,
+			ValidArgsFunction: cobra.NoFileCompletions,
+			RunE:              c.executeClean,
+		},
+		&cobra.Command{
+			Use:               "status",
+			Short:             "Show cache status",
+			Args:              cobra.NoArgs,
+			ValidArgsFunction: cobra.NoFileCompletions,
+			Run:               c.executeStatus,
+		},
+	)
 
-	// TODO: add trim command?
+	c.cmd = cacheCmd
+
+	return c
 }
 
-func (e *Executor) executeCleanCache(_ *cobra.Command, _ []string) error {
+func (*cacheCommand) executeClean(_ *cobra.Command, _ []string) error {
 	cacheDir := cache.DefaultDir()
+
 	if err := os.RemoveAll(cacheDir); err != nil {
 		return fmt.Errorf("failed to remove dir %s: %w", cacheDir, err)
 	}
@@ -50,13 +60,14 @@ func (e *Executor) executeCleanCache(_ *cobra.Command, _ []string) error {
 	return nil
 }
 
-func (e *Executor) executeCacheStatus(_ *cobra.Command, _ []string) {
+func (*cacheCommand) executeStatus(_ *cobra.Command, _ []string) {
 	cacheDir := cache.DefaultDir()
-	fmt.Fprintf(logutils.StdOut, "Dir: %s\n", cacheDir)
+
+	_, _ = fmt.Fprintf(logutils.StdOut, "Dir: %s\n", cacheDir)
 
 	cacheSizeBytes, err := dirSizeBytes(cacheDir)
 	if err == nil {
-		fmt.Fprintf(logutils.StdOut, "Size: %s\n", fsutils.PrettifyBytesCount(cacheSizeBytes))
+		_, _ = fmt.Fprintf(logutils.StdOut, "Size: %s\n", fsutils.PrettifyBytesCount(cacheSizeBytes))
 	}
 }
 
