@@ -1,174 +1,167 @@
 package commands
 
 import (
+	"errors"
 	"fmt"
 	"os"
-	"runtime"
-	"runtime/pprof"
-	"runtime/trace"
-	"strconv"
+	"slices"
 
+	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 
-	"github.com/golangci/golangci-lint/pkg/config"
-	"github.com/golangci/golangci-lint/pkg/exitcodes"
 	"github.com/golangci/golangci-lint/pkg/logutils"
 )
 
-const (
-	// envHelpRun value: "1".
-	envHelpRun        = "HELP_RUN"
-	envMemProfileRate = "GL_MEM_PROFILE_RATE"
-)
-
-func (e *Executor) persistentPreRun(_ *cobra.Command, _ []string) error {
-	if e.cfg.Run.PrintVersion {
-		_ = printVersion(logutils.StdOut, e.buildInfo)
-		os.Exit(exitcodes.Success) // a return nil is not enough to stop the process because we are inside the `preRun`.
-	}
-
-	runtime.GOMAXPROCS(e.cfg.Run.Concurrency)
-
-	if e.cfg.Run.CPUProfilePath != "" {
-		f, err := os.Create(e.cfg.Run.CPUProfilePath)
-		if err != nil {
-			return fmt.Errorf("can't create file %s: %w", e.cfg.Run.CPUProfilePath, err)
-		}
-		if err := pprof.StartCPUProfile(f); err != nil {
-			return fmt.Errorf("can't start CPU profiling: %w", err)
-		}
-	}
-
-	if e.cfg.Run.MemProfilePath != "" {
-		if rate := os.Getenv(envMemProfileRate); rate != "" {
-			runtime.MemProfileRate, _ = strconv.Atoi(rate)
-		}
-	}
-
-	if e.cfg.Run.TracePath != "" {
-		f, err := os.Create(e.cfg.Run.TracePath)
-		if err != nil {
-			return fmt.Errorf("can't create file %s: %w", e.cfg.Run.TracePath, err)
-		}
-		if err = trace.Start(f); err != nil {
-			return fmt.Errorf("can't start tracing: %w", err)
-		}
-	}
-
-	return nil
+func Execute(info BuildInfo) error {
+	return newRootCommand(info).Execute()
 }
 
-func (e *Executor) persistentPostRun(_ *cobra.Command, _ []string) error {
-	if e.cfg.Run.CPUProfilePath != "" {
-		pprof.StopCPUProfile()
-	}
+type rootOptions struct {
+	PrintVersion bool // Flag only.
 
-	if e.cfg.Run.MemProfilePath != "" {
-		f, err := os.Create(e.cfg.Run.MemProfilePath)
-		if err != nil {
-			return fmt.Errorf("can't create file %s: %w", e.cfg.Run.MemProfilePath, err)
-		}
-
-		var ms runtime.MemStats
-		runtime.ReadMemStats(&ms)
-		printMemStats(&ms, e.log)
-
-		if err := pprof.WriteHeapProfile(f); err != nil {
-			return fmt.Errorf("cCan't write heap profile: %w", err)
-		}
-		_ = f.Close()
-	}
-
-	if e.cfg.Run.TracePath != "" {
-		trace.Stop()
-	}
-
-	os.Exit(e.exitCode)
-
-	return nil
+	Verbose bool   // Flag only.
+	Color   string // Flag only.
 }
 
-func printMemStats(ms *runtime.MemStats, logger logutils.Log) {
-	logger.Infof("Mem stats: alloc=%s total_alloc=%s sys=%s "+
-		"heap_alloc=%s heap_sys=%s heap_idle=%s heap_released=%s heap_in_use=%s "+
-		"stack_in_use=%s stack_sys=%s "+
-		"mspan_sys=%s mcache_sys=%s buck_hash_sys=%s gc_sys=%s other_sys=%s "+
-		"mallocs_n=%d frees_n=%d heap_objects_n=%d gc_cpu_fraction=%.2f",
-		formatMemory(ms.Alloc), formatMemory(ms.TotalAlloc), formatMemory(ms.Sys),
-		formatMemory(ms.HeapAlloc), formatMemory(ms.HeapSys),
-		formatMemory(ms.HeapIdle), formatMemory(ms.HeapReleased), formatMemory(ms.HeapInuse),
-		formatMemory(ms.StackInuse), formatMemory(ms.StackSys),
-		formatMemory(ms.MSpanSys), formatMemory(ms.MCacheSys), formatMemory(ms.BuckHashSys),
-		formatMemory(ms.GCSys), formatMemory(ms.OtherSys),
-		ms.Mallocs, ms.Frees, ms.HeapObjects, ms.GCCPUFraction)
+type rootCommand struct {
+	cmd  *cobra.Command
+	opts rootOptions
+
+	log logutils.Log
 }
 
-func formatMemory(memBytes uint64) string {
-	const Kb = 1024
-	const Mb = Kb * 1024
+func newRootCommand(info BuildInfo) *rootCommand {
+	c := &rootCommand{}
 
-	if memBytes < Kb {
-		return fmt.Sprintf("%db", memBytes)
-	}
-	if memBytes < Mb {
-		return fmt.Sprintf("%dkb", memBytes/Kb)
-	}
-	return fmt.Sprintf("%dmb", memBytes/Mb)
-}
-
-func getDefaultConcurrency() int {
-	if os.Getenv(envHelpRun) == "1" {
-		// Make stable concurrency for README help generating builds.
-		const prettyConcurrency = 8
-		return prettyConcurrency
-	}
-
-	return runtime.NumCPU()
-}
-
-func (e *Executor) initRoot() {
 	rootCmd := &cobra.Command{
 		Use:   "golangci-lint",
 		Short: "golangci-lint is a smart linters runner.",
 		Long:  `Smart, fast linters runner.`,
 		Args:  cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, _ []string) error {
+			if c.opts.PrintVersion {
+				_ = printVersion(logutils.StdOut, info)
+				return nil
+			}
+
 			return cmd.Help()
 		},
-		PersistentPreRunE:  e.persistentPreRun,
-		PersistentPostRunE: e.persistentPostRun,
 	}
 
-	initRootFlagSet(rootCmd.PersistentFlags(), e.cfg, e.needVersionOption())
-	e.rootCmd = rootCmd
+	fs := rootCmd.Flags()
+	fs.BoolVar(&c.opts.PrintVersion, "version", false, color.GreenString("Print version"))
+
+	setupRootPersistentFlags(rootCmd.PersistentFlags(), &c.opts)
+
+	log := logutils.NewStderrLog(logutils.DebugKeyEmpty)
+
+	// Each command uses a dedicated configuration structure to avoid side effects of bindings.
+	rootCmd.AddCommand(
+		newLintersCommand(log).cmd,
+		newRunCommand(log, info).cmd,
+		newCacheCommand().cmd,
+		newConfigCommand(log, info).cmd,
+		newVersionCommand(info).cmd,
+		newCustomCommand(log).cmd,
+	)
+
+	rootCmd.SetHelpCommand(newHelpCommand(log).cmd)
+
+	c.log = log
+	c.cmd = rootCmd
+
+	return c
 }
 
-func (e *Executor) needVersionOption() bool {
-	return e.buildInfo.Date != ""
-}
-
-func initRootFlagSet(fs *pflag.FlagSet, cfg *config.Config, needVersionOption bool) {
-	fs.BoolVarP(&cfg.Run.IsVerbose, "verbose", "v", false, wh("Verbose output"))
-
-	var silent bool
-	fs.BoolVarP(&silent, "silent", "s", false, wh("Disables congrats outputs"))
-	if err := fs.MarkHidden("silent"); err != nil {
-		panic(err)
-	}
-	err := fs.MarkDeprecated("silent",
-		"now golangci-lint by default is silent: it doesn't print Congrats message")
+func (c *rootCommand) Execute() error {
+	err := setupLogger(c.log)
 	if err != nil {
-		panic(err)
+		return err
 	}
 
-	fs.StringVar(&cfg.Run.CPUProfilePath, "cpu-profile-path", "", wh("Path to CPU profile output file"))
-	fs.StringVar(&cfg.Run.MemProfilePath, "mem-profile-path", "", wh("Path to memory profile output file"))
-	fs.StringVar(&cfg.Run.TracePath, "trace-path", "", wh("Path to trace output file"))
-	fs.IntVarP(&cfg.Run.Concurrency, "concurrency", "j", getDefaultConcurrency(), wh("Concurrency (default NumCPU)"))
-	if needVersionOption {
-		fs.BoolVar(&cfg.Run.PrintVersion, "version", false, wh("Print version"))
+	return c.cmd.Execute()
+}
+
+func setupRootPersistentFlags(fs *pflag.FlagSet, opts *rootOptions) {
+	fs.BoolP("help", "h", false, color.GreenString("Help for a command"))
+	fs.BoolVarP(&opts.Verbose, "verbose", "v", false, color.GreenString("Verbose output"))
+	fs.StringVar(&opts.Color, "color", "auto", color.GreenString("Use color when printing; can be 'always', 'auto', or 'never'"))
+}
+
+func setupLogger(logger logutils.Log) error {
+	opts, err := forceRootParsePersistentFlags()
+	if err != nil && !errors.Is(err, pflag.ErrHelp) {
+		return err
 	}
 
-	fs.StringVar(&cfg.Output.Color, "color", "auto", wh("Use color when printing; can be 'always', 'auto', or 'never'"))
+	if opts == nil {
+		return nil
+	}
+
+	logutils.SetupVerboseLog(logger, opts.Verbose)
+
+	switch opts.Color {
+	case "always":
+		color.NoColor = false
+	case "never":
+		color.NoColor = true
+	case "auto":
+		// nothing
+	default:
+		logger.Fatalf("invalid value %q for --color; must be 'always', 'auto', or 'never'", opts.Color)
+	}
+
+	return nil
+}
+
+func forceRootParsePersistentFlags() (*rootOptions, error) {
+	// We use another pflag.FlagSet here to not set `changed` flag on cmd.Flags() options.
+	// Otherwise, string slice options will be duplicated.
+	fs := pflag.NewFlagSet("config flag set", pflag.ContinueOnError)
+
+	// Ignore unknown flags because we will parse the command flags later.
+	fs.ParseErrorsWhitelist = pflag.ParseErrorsWhitelist{UnknownFlags: true}
+
+	opts := &rootOptions{}
+
+	// Don't do `fs.AddFlagSet(cmd.Flags())` because it shares flags representations:
+	// `changed` variable inside string slice vars will be shared.
+	// Use another config variable here,
+	// to not affect main parsing by this parsing of only config option.
+	setupRootPersistentFlags(fs, opts)
+
+	fs.Usage = func() {} // otherwise, help text will be printed twice
+
+	if err := fs.Parse(safeArgs(fs, os.Args)); err != nil {
+		if errors.Is(err, pflag.ErrHelp) {
+			return nil, err
+		}
+
+		return nil, fmt.Errorf("can't parse args: %w", err)
+	}
+
+	return opts, nil
+}
+
+// Shorthands are a problem because pflag, with UnknownFlags, will try to parse all the letters as options.
+// A shorthand can aggregate several letters (ex `ps -aux`)
+// The function replaces non-supported shorthands by a dumb flag.
+func safeArgs(fs *pflag.FlagSet, args []string) []string {
+	var shorthands []string
+	fs.VisitAll(func(flag *pflag.Flag) {
+		shorthands = append(shorthands, flag.Shorthand)
+	})
+
+	var cleanArgs []string
+	for _, arg := range args {
+		if len(arg) > 1 && arg[0] == '-' && arg[1] != '-' && !slices.Contains(shorthands, string(arg[1])) {
+			cleanArgs = append(cleanArgs, "--potato")
+			continue
+		}
+
+		cleanArgs = append(cleanArgs, arg)
+	}
+
+	return cleanArgs
 }
