@@ -8,76 +8,95 @@ import (
 	"runtime/debug"
 	"strings"
 
+	"github.com/fatih/color"
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
-
-	"github.com/golangci/golangci-lint/pkg/config"
 )
+
+type BuildInfo struct {
+	GoVersion string `json:"goVersion"`
+	Version   string `json:"version"`
+	Commit    string `json:"commit"`
+	Date      string `json:"date"`
+}
+
+func (b BuildInfo) String() string {
+	return fmt.Sprintf("golangci-lint has version %s built with %s from %s on %s",
+		b.Version, b.GoVersion, b.Commit, b.Date)
+}
 
 type versionInfo struct {
 	Info      BuildInfo
 	BuildInfo *debug.BuildInfo
 }
 
-func (e *Executor) initVersionConfiguration(cmd *cobra.Command) {
-	fs := cmd.Flags()
-	fs.SortFlags = false // sort them as they are defined here
-	initVersionFlagSet(fs, e.cfg)
+type versionOptions struct {
+	Format string
+	Debug  bool
 }
 
-func initVersionFlagSet(fs *pflag.FlagSet, cfg *config.Config) {
-	// Version config
-	vc := &cfg.Version
-	fs.StringVar(&vc.Format, "format", "", wh("The version's format can be: 'short', 'json'"))
-	fs.BoolVar(&vc.Debug, "debug", false, wh("Add build information"))
+type versionCommand struct {
+	cmd  *cobra.Command
+	opts versionOptions
+
+	info BuildInfo
 }
 
-func (e *Executor) initVersion() {
+func newVersionCommand(info BuildInfo) *versionCommand {
+	c := &versionCommand{info: info}
+
 	versionCmd := &cobra.Command{
 		Use:               "version",
 		Short:             "Version",
 		Args:              cobra.NoArgs,
 		ValidArgsFunction: cobra.NoFileCompletions,
-		RunE: func(cmd *cobra.Command, _ []string) error {
-			if e.cfg.Version.Debug {
-				info, ok := debug.ReadBuildInfo()
-				if !ok {
-					return nil
-				}
-
-				switch strings.ToLower(e.cfg.Version.Format) {
-				case "json":
-					return json.NewEncoder(os.Stdout).Encode(versionInfo{
-						Info:      e.buildInfo,
-						BuildInfo: info,
-					})
-
-				default:
-					fmt.Println(info.String())
-					return printVersion(os.Stdout, e.buildInfo)
-				}
-			}
-
-			switch strings.ToLower(e.cfg.Version.Format) {
-			case "short":
-				fmt.Println(e.buildInfo.Version)
-				return nil
-
-			case "json":
-				return json.NewEncoder(os.Stdout).Encode(e.buildInfo)
-
-			default:
-				return printVersion(os.Stdout, e.buildInfo)
-			}
-		},
+		RunE:              c.execute,
 	}
 
-	e.rootCmd.AddCommand(versionCmd)
-	e.initVersionConfiguration(versionCmd)
+	fs := versionCmd.Flags()
+	fs.SortFlags = false // sort them as they are defined here
+
+	fs.StringVar(&c.opts.Format, "format", "", color.GreenString("The version's format can be: 'short', 'json'"))
+	fs.BoolVar(&c.opts.Debug, "debug", false, color.GreenString("Add build information"))
+
+	c.cmd = versionCmd
+
+	return c
 }
 
-func printVersion(w io.Writer, buildInfo BuildInfo) error {
-	_, err := fmt.Fprintf(w, "golangci-lint has version %s built with %s from %s on %s\n",
-		buildInfo.Version, buildInfo.GoVersion, buildInfo.Commit, buildInfo.Date)
+func (c *versionCommand) execute(_ *cobra.Command, _ []string) error {
+	if c.opts.Debug {
+		info, ok := debug.ReadBuildInfo()
+		if !ok {
+			return nil
+		}
+
+		switch strings.ToLower(c.opts.Format) {
+		case "json":
+			return json.NewEncoder(os.Stdout).Encode(versionInfo{
+				Info:      c.info,
+				BuildInfo: info,
+			})
+
+		default:
+			fmt.Println(info.String())
+			return printVersion(os.Stdout, c.info)
+		}
+	}
+
+	switch strings.ToLower(c.opts.Format) {
+	case "short":
+		fmt.Println(c.info.Version)
+		return nil
+
+	case "json":
+		return json.NewEncoder(os.Stdout).Encode(c.info)
+
+	default:
+		return printVersion(os.Stdout, c.info)
+	}
+}
+
+func printVersion(w io.Writer, info BuildInfo) error {
+	_, err := fmt.Fprintln(w, info.String())
 	return err
 }
