@@ -3,6 +3,7 @@ package config
 import (
 	"encoding"
 	"errors"
+	"fmt"
 	"runtime"
 
 	"gopkg.in/yaml.v3"
@@ -20,6 +21,9 @@ var defaultLintersSettings = LintersSettings{
 	},
 	Dogsled: DogsledSettings{
 		MaxBlankIdentifiers: 2,
+	},
+	Dupl: DuplSettings{
+		Threshold: 150,
 	},
 	ErrorLint: ErrorLintSettings{
 		Errorf:      true,
@@ -43,11 +47,25 @@ var defaultLintersSettings = LintersSettings{
 		Sections:      []string{"standard", "default"},
 		SkipGenerated: true,
 	},
+	GoChecksumType: GoChecksumTypeSettings{
+		DefaultSignifiesExhaustive: true,
+	},
 	Gocognit: GocognitSettings{
 		MinComplexity: 30,
 	},
+	Goconst: GoConstSettings{
+		MatchWithConstants:  true,
+		MinStringLen:        3,
+		MinOccurrencesCount: 3,
+		NumberMin:           3,
+		NumberMax:           3,
+		IgnoreCalls:         true,
+	},
 	Gocritic: GoCriticSettings{
 		SettingsPerCheck: map[string]GoCriticCheckSettings{},
+	},
+	Gocyclo: GoCycloSettings{
+		MinComplexity: 30,
 	},
 	Godox: GodoxSettings{
 		Keywords: []string{},
@@ -55,6 +73,9 @@ var defaultLintersSettings = LintersSettings{
 	Godot: GodotSettings{
 		Scope:  "declarations",
 		Period: true,
+	},
+	Gofmt: GoFmtSettings{
+		Simplify: true,
 	},
 	Gofumpt: GofumptSettings{
 		LangVersion: "",
@@ -70,9 +91,8 @@ var defaultLintersSettings = LintersSettings{
 		IgnoreTests:     true,
 		WatchForScripts: []string{"Han"},
 	},
-	Ifshort: IfshortSettings{
-		MaxDeclLines: 1,
-		MaxDeclChars: 30,
+	Inamedparam: INamedParamSettings{
+		SkipSingleParam: false,
 	},
 	InterfaceBloat: InterfaceBloatSettings{
 		Max: 10,
@@ -85,6 +105,7 @@ var defaultLintersSettings = LintersSettings{
 		Kitlog:           true,
 		Klog:             true,
 		Logr:             true,
+		Slog:             true,
 		Zap:              true,
 		RequireStringKey: false,
 		NoPrintfLike:     false,
@@ -105,10 +126,16 @@ var defaultLintersSettings = LintersSettings{
 		AllowUnused:        false,
 	},
 	PerfSprint: PerfSprintSettings{
+		IntegerFormat: true,
 		IntConversion: true,
+		ErrorFormat:   true,
 		ErrError:      false,
 		ErrorF:        true,
+		StringFormat:  true,
 		SprintF1:      true,
+		StrConcat:     true,
+		BoolFormat:    true,
+		HexFormat:     true,
 	},
 	Prealloc: PreallocSettings{
 		Simple:     true,
@@ -120,11 +147,15 @@ var defaultLintersSettings = LintersSettings{
 		Qualified: false,
 	},
 	SlogLint: SlogLintSettings{
+		NoMixedArgs:    true,
 		KVOnly:         false,
 		AttrOnly:       false,
-		ContextOnly:    false,
+		NoGlobal:       "",
+		Context:        "",
+		StaticMsg:      false,
 		NoRawKeys:      false,
 		KeyNamingCase:  "",
+		ForbiddenKeys:  nil,
 		ArgsOnSepLines: false,
 	},
 	TagAlign: TagAlignSettings{
@@ -143,7 +174,6 @@ var defaultLintersSettings = LintersSettings{
 	Unused: UnusedSettings{
 		FieldWritesAreUses:     true,
 		PostStatementsAreReads: false,
-		ExportedIsUsed:         true,
 		ExportedFieldsAreUsed:  true,
 		ParametersAreUsed:      true,
 		LocalVariablesAreUsed:  true,
@@ -152,6 +182,15 @@ var defaultLintersSettings = LintersSettings{
 	UseStdlibVars: UseStdlibVarsSettings{
 		HTTPMethod:     true,
 		HTTPStatusCode: true,
+	},
+	UseTesting: UseTestingSettings{
+		ContextBackground: true,
+		ContextTodo:       true,
+		OSChdir:           true,
+		OSMkdirTemp:       true,
+		OSSetenv:          true,
+		OSTempDir:         false,
+		OSCreateTemp:      true,
 	},
 	Varnamelen: VarnamelenSettings{
 		MaxDistance:   5,
@@ -175,89 +214,106 @@ var defaultLintersSettings = LintersSettings{
 }
 
 type LintersSettings struct {
-	Asasalint        AsasalintSettings
-	BiDiChk          BiDiChkSettings
-	Cyclop           Cyclop
-	Decorder         DecorderSettings
-	Depguard         DepGuardSettings
-	Dogsled          DogsledSettings
-	Dupl             DuplSettings
-	DupWord          DupWordSettings
-	Errcheck         ErrcheckSettings
-	ErrChkJSON       ErrChkJSONSettings
-	ErrorLint        ErrorLintSettings
-	Exhaustive       ExhaustiveSettings
-	ExhaustiveStruct ExhaustiveStructSettings
-	Exhaustruct      ExhaustructSettings
-	Forbidigo        ForbidigoSettings
-	Funlen           FunlenSettings
-	Gci              GciSettings
-	GinkgoLinter     GinkgoLinterSettings
-	Gocognit         GocognitSettings
-	Goconst          GoConstSettings
-	Gocritic         GoCriticSettings
-	Gocyclo          GoCycloSettings
-	Godot            GodotSettings
-	Godox            GodoxSettings
-	Gofmt            GoFmtSettings
-	Gofumpt          GofumptSettings
-	Goheader         GoHeaderSettings
-	Goimports        GoImportsSettings
-	Golint           GoLintSettings
-	Gomnd            GoMndSettings
-	GoModDirectives  GoModDirectivesSettings
-	Gomodguard       GoModGuardSettings
-	Gosec            GoSecSettings
-	Gosimple         StaticCheckSettings
-	Gosmopolitan     GosmopolitanSettings
-	Govet            GovetSettings
-	Grouper          GrouperSettings
-	Ifshort          IfshortSettings
-	ImportAs         ImportAsSettings
-	InterfaceBloat   InterfaceBloatSettings
-	Ireturn          IreturnSettings
-	Lll              LllSettings
-	LoggerCheck      LoggerCheckSettings
-	MaintIdx         MaintIdxSettings
-	Makezero         MakezeroSettings
-	Maligned         MalignedSettings
-	Misspell         MisspellSettings
-	MustTag          MustTagSettings
-	Nakedret         NakedretSettings
-	Nestif           NestifSettings
-	NilNil           NilNilSettings
-	Nlreturn         NlreturnSettings
-	NoLintLint       NoLintLintSettings
-	NoNamedReturns   NoNamedReturnsSettings
-	ParallelTest     ParallelTestSettings
-	PerfSprint       PerfSprintSettings
-	Prealloc         PreallocSettings
-	Predeclared      PredeclaredSettings
-	Promlinter       PromlinterSettings
-	ProtoGetter      ProtoGetterSettings
-	Reassign         ReassignSettings
-	Revive           ReviveSettings
-	RowsErrCheck     RowsErrCheckSettings
-	SlogLint         SlogLintSettings
-	Staticcheck      StaticCheckSettings
-	Structcheck      StructCheckSettings
-	Stylecheck       StaticCheckSettings
-	TagAlign         TagAlignSettings
-	Tagliatelle      TagliatelleSettings
-	Testifylint      TestifylintSettings
-	Tenv             TenvSettings
-	Testpackage      TestpackageSettings
-	Thelper          ThelperSettings
-	Unparam          UnparamSettings
-	Unused           UnusedSettings
-	UseStdlibVars    UseStdlibVarsSettings
-	Varcheck         VarCheckSettings
-	Varnamelen       VarnamelenSettings
-	Whitespace       WhitespaceSettings
-	Wrapcheck        WrapcheckSettings
-	WSL              WSLSettings
+	Asasalint       AsasalintSettings
+	BiDiChk         BiDiChkSettings
+	CopyLoopVar     CopyLoopVarSettings
+	Cyclop          Cyclop
+	Decorder        DecorderSettings
+	Depguard        DepGuardSettings
+	Dogsled         DogsledSettings
+	Dupl            DuplSettings
+	DupWord         DupWordSettings
+	Errcheck        ErrcheckSettings
+	ErrChkJSON      ErrChkJSONSettings
+	ErrorLint       ErrorLintSettings
+	Exhaustive      ExhaustiveSettings
+	Exhaustruct     ExhaustructSettings
+	Fatcontext      FatcontextSettings
+	Forbidigo       ForbidigoSettings
+	Funlen          FunlenSettings
+	Gci             GciSettings
+	GinkgoLinter    GinkgoLinterSettings
+	Gocognit        GocognitSettings
+	GoChecksumType  GoChecksumTypeSettings
+	Goconst         GoConstSettings
+	Gocritic        GoCriticSettings
+	Gocyclo         GoCycloSettings
+	Godot           GodotSettings
+	Godox           GodoxSettings
+	Gofmt           GoFmtSettings
+	Gofumpt         GofumptSettings
+	Goheader        GoHeaderSettings
+	Goimports       GoImportsSettings
+	GoModDirectives GoModDirectivesSettings
+	Gomodguard      GoModGuardSettings
+	Gosec           GoSecSettings
+	Gosimple        StaticCheckSettings
+	Gosmopolitan    GosmopolitanSettings
+	Govet           GovetSettings
+	Grouper         GrouperSettings
+	Iface           IfaceSettings
+	ImportAs        ImportAsSettings
+	Inamedparam     INamedParamSettings
+	InterfaceBloat  InterfaceBloatSettings
+	Ireturn         IreturnSettings
+	Lll             LllSettings
+	LoggerCheck     LoggerCheckSettings
+	MaintIdx        MaintIdxSettings
+	Makezero        MakezeroSettings
+	Misspell        MisspellSettings
+	Mnd             MndSettings
+	MustTag         MustTagSettings
+	Nakedret        NakedretSettings
+	Nestif          NestifSettings
+	NilNil          NilNilSettings
+	Nlreturn        NlreturnSettings
+	NoLintLint      NoLintLintSettings
+	NoNamedReturns  NoNamedReturnsSettings
+	ParallelTest    ParallelTestSettings
+	PerfSprint      PerfSprintSettings
+	Prealloc        PreallocSettings
+	Predeclared     PredeclaredSettings
+	Promlinter      PromlinterSettings
+	ProtoGetter     ProtoGetterSettings
+	Reassign        ReassignSettings
+	Recvcheck       RecvcheckSettings
+	Revive          ReviveSettings
+	RowsErrCheck    RowsErrCheckSettings
+	SlogLint        SlogLintSettings
+	Spancheck       SpancheckSettings
+	Staticcheck     StaticCheckSettings
+	Stylecheck      StaticCheckSettings
+	TagAlign        TagAlignSettings
+	Tagliatelle     TagliatelleSettings
+	Tenv            TenvSettings
+	Testifylint     TestifylintSettings
+	Testpackage     TestpackageSettings
+	Thelper         ThelperSettings
+	Unconvert       UnconvertSettings
+	Unparam         UnparamSettings
+	Unused          UnusedSettings
+	UseStdlibVars   UseStdlibVarsSettings
+	UseTesting      UseTestingSettings
+	Varnamelen      VarnamelenSettings
+	Whitespace      WhitespaceSettings
+	Wrapcheck       WrapcheckSettings
+	WSL             WSLSettings
 
 	Custom map[string]CustomLinterSettings
+}
+
+func (s *LintersSettings) Validate() error {
+	if err := s.Govet.Validate(); err != nil {
+		return err
+	}
+
+	for name, settings := range s.Custom {
+		if err := settings.Validate(); err != nil {
+			return fmt.Errorf("custom linter %q: %w", name, err)
+		}
+	}
+
+	return nil
 }
 
 type AsasalintSettings struct {
@@ -276,6 +332,13 @@ type BiDiChkSettings struct {
 	RightToLeftIsolate       bool `mapstructure:"right-to-left-isolate"`
 	FirstStrongIsolate       bool `mapstructure:"first-strong-isolate"`
 	PopDirectionalIsolate    bool `mapstructure:"pop-directional-isolate"`
+}
+
+type CopyLoopVarSettings struct {
+	CheckAlias bool `mapstructure:"check-alias"`
+
+	// Deprecated: use CheckAlias
+	IgnoreAlias *bool `mapstructure:"ignore-alias"`
 }
 
 type Cyclop struct {
@@ -328,11 +391,13 @@ type ErrcheckSettings struct {
 	DisableDefaultExclusions bool     `mapstructure:"disable-default-exclusions"`
 	CheckTypeAssertions      bool     `mapstructure:"check-type-assertions"`
 	CheckAssignToBlank       bool     `mapstructure:"check-blank"`
-	Ignore                   string   `mapstructure:"ignore"`
 	ExcludeFunctions         []string `mapstructure:"exclude-functions"`
 
 	// Deprecated: use ExcludeFunctions instead
 	Exclude string `mapstructure:"exclude"`
+
+	// Deprecated: use ExcludeFunctions instead
+	Ignore string `mapstructure:"ignore"`
 }
 
 type ErrChkJSONSettings struct {
@@ -341,10 +406,17 @@ type ErrChkJSONSettings struct {
 }
 
 type ErrorLintSettings struct {
-	Errorf      bool `mapstructure:"errorf"`
-	ErrorfMulti bool `mapstructure:"errorf-multi"`
-	Asserts     bool `mapstructure:"asserts"`
-	Comparison  bool `mapstructure:"comparison"`
+	Errorf                bool                 `mapstructure:"errorf"`
+	ErrorfMulti           bool                 `mapstructure:"errorf-multi"`
+	Asserts               bool                 `mapstructure:"asserts"`
+	Comparison            bool                 `mapstructure:"comparison"`
+	AllowedErrors         []ErrorLintAllowPair `mapstructure:"allowed-errors"`
+	AllowedErrorsWildcard []ErrorLintAllowPair `mapstructure:"allowed-errors-wildcard"`
+}
+
+type ErrorLintAllowPair struct {
+	Err string `mapstructure:"err"`
+	Fun string `mapstructure:"fun"`
 }
 
 type ExhaustiveSettings struct {
@@ -359,13 +431,13 @@ type ExhaustiveSettings struct {
 	DefaultCaseRequired        bool     `mapstructure:"default-case-required"`
 }
 
-type ExhaustiveStructSettings struct {
-	StructPatterns []string `mapstructure:"struct-patterns"`
-}
-
 type ExhaustructSettings struct {
 	Include []string `mapstructure:"include"`
 	Exclude []string `mapstructure:"exclude"`
+}
+
+type FatcontextSettings struct {
+	CheckStructPointers bool `mapstructure:"check-struct-pointers"`
 }
 
 type ForbidigoSettings struct {
@@ -417,10 +489,15 @@ type FunlenSettings struct {
 }
 
 type GciSettings struct {
-	LocalPrefixes string   `mapstructure:"local-prefixes"` // Deprecated
-	Sections      []string `mapstructure:"sections"`
-	SkipGenerated bool     `mapstructure:"skip-generated"`
-	CustomOrder   bool     `mapstructure:"custom-order"`
+	Sections         []string `mapstructure:"sections"`
+	NoInlineComments bool     `mapstructure:"no-inline-comments"`
+	NoPrefixComments bool     `mapstructure:"no-prefix-comments"`
+	SkipGenerated    bool     `mapstructure:"skip-generated"`
+	CustomOrder      bool     `mapstructure:"custom-order"`
+	NoLexOrder       bool     `mapstructure:"no-lex-order"`
+
+	// Deprecated: use Sections instead.
+	LocalPrefixes string `mapstructure:"local-prefixes"`
 }
 
 type GinkgoLinterSettings struct {
@@ -432,6 +509,15 @@ type GinkgoLinterSettings struct {
 	SuppressTypeCompareWarning bool `mapstructure:"suppress-type-compare-assertion"`
 	ForbidFocusContainer       bool `mapstructure:"forbid-focus-container"`
 	AllowHaveLenZero           bool `mapstructure:"allow-havelen-zero"`
+	ForceExpectTo              bool `mapstructure:"force-expect-to"`
+	ValidateAsyncIntervals     bool `mapstructure:"validate-async-intervals"`
+	ForbidSpecPollution        bool `mapstructure:"forbid-spec-pollution"`
+	ForceSucceedForFuncs       bool `mapstructure:"force-succeed"`
+}
+
+type GoChecksumTypeSettings struct {
+	DefaultSignifiesExhaustive bool `mapstructure:"default-signifies-exhaustive"`
+	IncludeSharedInterfaces    bool `mapstructure:"include-shared-interfaces"`
 }
 
 type GocognitSettings struct {
@@ -452,7 +538,9 @@ type GoConstSettings struct {
 
 type GoCriticSettings struct {
 	Go               string                           `mapstructure:"-"`
+	DisableAll       bool                             `mapstructure:"disable-all"`
 	EnabledChecks    []string                         `mapstructure:"enabled-checks"`
+	EnableAll        bool                             `mapstructure:"enable-all"`
 	DisabledChecks   []string                         `mapstructure:"disabled-checks"`
 	EnabledTags      []string                         `mapstructure:"enabled-tags"`
 	DisabledTags     []string                         `mapstructure:"disabled-tags"`
@@ -471,8 +559,8 @@ type GodotSettings struct {
 	Capital bool     `mapstructure:"capital"`
 	Period  bool     `mapstructure:"period"`
 
-	// Deprecated: use `Scope` instead
-	CheckAll bool `mapstructure:"check-all"`
+	// Deprecated: use Scope instead
+	CheckAll *bool `mapstructure:"check-all"`
 }
 
 type GodoxSettings struct {
@@ -507,23 +595,16 @@ type GoImportsSettings struct {
 	LocalPrefixes string `mapstructure:"local-prefixes"`
 }
 
-type GoLintSettings struct {
-	MinConfidence float64 `mapstructure:"min-confidence"`
-}
-
-type GoMndSettings struct {
-	Settings         map[string]map[string]any // Deprecated
-	Checks           []string                  `mapstructure:"checks"`
-	IgnoredNumbers   []string                  `mapstructure:"ignored-numbers"`
-	IgnoredFiles     []string                  `mapstructure:"ignored-files"`
-	IgnoredFunctions []string                  `mapstructure:"ignored-functions"`
-}
-
 type GoModDirectivesSettings struct {
 	ReplaceAllowList          []string `mapstructure:"replace-allow-list"`
 	ReplaceLocal              bool     `mapstructure:"replace-local"`
 	ExcludeForbidden          bool     `mapstructure:"exclude-forbidden"`
 	RetractAllowNoExplanation bool     `mapstructure:"retract-allow-no-explanation"`
+	ToolchainForbidden        bool     `mapstructure:"toolchain-forbidden"`
+	ToolchainPattern          string   `mapstructure:"toolchain-pattern"`
+	ToolForbidden             bool     `mapstructure:"tool-forbidden"`
+	GoDebugForbidden          bool     `mapstructure:"go-debug-forbidden"`
+	GoVersionPattern          string   `mapstructure:"go-version-pattern"`
 }
 
 type GoModGuardSettings struct {
@@ -562,25 +643,28 @@ type GosmopolitanSettings struct {
 }
 
 type GovetSettings struct {
-	Go             string `mapstructure:"-"`
-	CheckShadowing bool   `mapstructure:"check-shadowing"`
-	Settings       map[string]map[string]any
+	Go string `mapstructure:"-"`
 
 	Enable     []string
 	Disable    []string
 	EnableAll  bool `mapstructure:"enable-all"`
 	DisableAll bool `mapstructure:"disable-all"`
+
+	Settings map[string]map[string]any
+
+	// Deprecated: the linter should be enabled inside Enable.
+	CheckShadowing *bool `mapstructure:"check-shadowing"`
 }
 
 func (cfg *GovetSettings) Validate() error {
 	if cfg.EnableAll && cfg.DisableAll {
-		return errors.New("enable-all and disable-all can't be combined")
+		return errors.New("govet: enable-all and disable-all can't be combined")
 	}
 	if cfg.EnableAll && len(cfg.Enable) != 0 {
-		return errors.New("enable-all and enable can't be combined")
+		return errors.New("govet: enable-all and enable can't be combined")
 	}
 	if cfg.DisableAll && len(cfg.Disable) != 0 {
-		return errors.New("disable-all and disable can't be combined")
+		return errors.New("govet: disable-all and disable can't be combined")
 	}
 	return nil
 }
@@ -596,9 +680,9 @@ type GrouperSettings struct {
 	VarRequireGrouping        bool `mapstructure:"var-require-grouping"`
 }
 
-type IfshortSettings struct {
-	MaxDeclLines int `mapstructure:"max-decl-lines"`
-	MaxDeclChars int `mapstructure:"max-decl-chars"`
+type IfaceSettings struct {
+	Enable   []string                  `mapstructure:"enable"`
+	Settings map[string]map[string]any `mapstructure:"settings"`
 }
 
 type ImportAsSettings struct {
@@ -610,6 +694,10 @@ type ImportAsSettings struct {
 type ImportAsAlias struct {
 	Pkg   string
 	Alias string
+}
+
+type INamedParamSettings struct {
+	SkipSingleParam bool `mapstructure:"skip-single-param"`
 }
 
 type InterfaceBloatSettings struct {
@@ -630,6 +718,7 @@ type LoggerCheckSettings struct {
 	Kitlog           bool     `mapstructure:"kitlog"`
 	Klog             bool     `mapstructure:"klog"`
 	Logr             bool     `mapstructure:"logr"`
+	Slog             bool     `mapstructure:"slog"`
 	Zap              bool     `mapstructure:"zap"`
 	RequireStringKey bool     `mapstructure:"require-string-key"`
 	NoPrintfLike     bool     `mapstructure:"no-printf-like"`
@@ -644,14 +733,17 @@ type MakezeroSettings struct {
 	Always bool
 }
 
-type MalignedSettings struct {
-	SuggestNewOrder bool `mapstructure:"suggest-new"`
+type MisspellSettings struct {
+	Mode       string               `mapstructure:"mode"`
+	Locale     string               `mapstructure:"locale"`
+	ExtraWords []MisspellExtraWords `mapstructure:"extra-words"`
+	// TODO(ldez): v2 the option must be renamed to `IgnoredRules`.
+	IgnoreWords []string `mapstructure:"ignore-words"`
 }
 
-type MisspellSettings struct {
-	Locale string
-	// TODO(ldez): v2 the options must be renamed to `IgnoredRules`.
-	IgnoreWords []string `mapstructure:"ignore-words"`
+type MisspellExtraWords struct {
+	Typo       string `mapstructure:"typo"`
+	Correction string `mapstructure:"correction"`
 }
 
 type MustTagSettings struct {
@@ -663,7 +755,7 @@ type MustTagSettings struct {
 }
 
 type NakedretSettings struct {
-	MaxFuncLines int `mapstructure:"max-func-lines"`
+	MaxFuncLines uint `mapstructure:"max-func-lines"`
 }
 
 type NestifSettings struct {
@@ -671,11 +763,19 @@ type NestifSettings struct {
 }
 
 type NilNilSettings struct {
-	CheckedTypes []string `mapstructure:"checked-types"`
+	DetectOpposite bool     `mapstructure:"detect-opposite"`
+	CheckedTypes   []string `mapstructure:"checked-types"`
 }
 
 type NlreturnSettings struct {
 	BlockSize int `mapstructure:"block-size"`
+}
+
+type MndSettings struct {
+	Checks           []string `mapstructure:"checks"`
+	IgnoredNumbers   []string `mapstructure:"ignored-numbers"`
+	IgnoredFiles     []string `mapstructure:"ignored-files"`
+	IgnoredFunctions []string `mapstructure:"ignored-functions"`
 }
 
 type NoLintLintSettings struct {
@@ -690,15 +790,25 @@ type NoNamedReturnsSettings struct {
 }
 
 type ParallelTestSettings struct {
-	IgnoreMissing         bool `mapstructure:"ignore-missing"`
-	IgnoreMissingSubtests bool `mapstructure:"ignore-missing-subtests"`
+	Go                    string `mapstructure:"-"`
+	IgnoreMissing         bool   `mapstructure:"ignore-missing"`
+	IgnoreMissingSubtests bool   `mapstructure:"ignore-missing-subtests"`
 }
 
 type PerfSprintSettings struct {
+	IntegerFormat bool `mapstructure:"integer-format"`
 	IntConversion bool `mapstructure:"int-conversion"`
-	ErrError      bool `mapstructure:"err-error"`
-	ErrorF        bool `mapstructure:"errorf"`
-	SprintF1      bool `mapstructure:"sprintf1"`
+
+	ErrorFormat bool `mapstructure:"error-format"`
+	ErrError    bool `mapstructure:"err-error"`
+	ErrorF      bool `mapstructure:"errorf"`
+
+	StringFormat bool `mapstructure:"string-format"`
+	SprintF1     bool `mapstructure:"sprintf1"`
+	StrConcat    bool `mapstructure:"strconcat"`
+
+	BoolFormat bool `mapstructure:"bool-format"`
+	HexFormat  bool `mapstructure:"hex-format"`
 }
 
 type PreallocSettings struct {
@@ -718,18 +828,25 @@ type PromlinterSettings struct {
 }
 
 type ProtoGetterSettings struct {
-	SkipGeneratedBy  []string `mapstructure:"skip-generated-by"`
-	SkipFiles        []string `mapstructure:"skip-files"`
-	SkipAnyGenerated bool     `mapstructure:"skip-any-generated"`
+	SkipGeneratedBy         []string `mapstructure:"skip-generated-by"`
+	SkipFiles               []string `mapstructure:"skip-files"`
+	SkipAnyGenerated        bool     `mapstructure:"skip-any-generated"`
+	ReplaceFirstArgInAppend bool     `mapstructure:"replace-first-arg-in-append"`
 }
 
 type ReassignSettings struct {
 	Patterns []string `mapstructure:"patterns"`
 }
 
+type RecvcheckSettings struct {
+	DisableBuiltin bool     `mapstructure:"disable-builtin"`
+	Exclusions     []string `mapstructure:"exclusions"`
+}
+
 type ReviveSettings struct {
-	MaxOpenFiles          int  `mapstructure:"max-open-files"`
-	IgnoreGeneratedHeader bool `mapstructure:"ignore-generated-header"`
+	Go                    string `mapstructure:"-"`
+	MaxOpenFiles          int    `mapstructure:"max-open-files"`
+	IgnoreGeneratedHeader bool   `mapstructure:"ignore-generated-header"`
 	Confidence            float64
 	Severity              string
 	EnableAllRules        bool `mapstructure:"enable-all-rules"`
@@ -738,6 +855,7 @@ type ReviveSettings struct {
 		Arguments []any
 		Severity  string
 		Disabled  bool
+		Exclude   []string
 	}
 	ErrorCode   int `mapstructure:"error-code"`
 	WarningCode int `mapstructure:"warning-code"`
@@ -752,30 +870,39 @@ type RowsErrCheckSettings struct {
 }
 
 type SlogLintSettings struct {
-	KVOnly         bool   `mapstructure:"kv-only"`
-	AttrOnly       bool   `mapstructure:"attr-only"`
-	ContextOnly    bool   `mapstructure:"context-only"`
-	NoRawKeys      bool   `mapstructure:"no-raw-keys"`
-	KeyNamingCase  string `mapstructure:"key-naming-case"`
-	ArgsOnSepLines bool   `mapstructure:"args-on-sep-lines"`
+	NoMixedArgs    bool     `mapstructure:"no-mixed-args"`
+	KVOnly         bool     `mapstructure:"kv-only"`
+	AttrOnly       bool     `mapstructure:"attr-only"`
+	NoGlobal       string   `mapstructure:"no-global"`
+	Context        string   `mapstructure:"context"`
+	StaticMsg      bool     `mapstructure:"static-msg"`
+	NoRawKeys      bool     `mapstructure:"no-raw-keys"`
+	KeyNamingCase  string   `mapstructure:"key-naming-case"`
+	ForbiddenKeys  []string `mapstructure:"forbidden-keys"`
+	ArgsOnSepLines bool     `mapstructure:"args-on-sep-lines"`
+
+	// Deprecated: use Context instead.
+	ContextOnly *bool `mapstructure:"context-only"`
+}
+
+type SpancheckSettings struct {
+	Checks                   []string `mapstructure:"checks"`
+	IgnoreCheckSignatures    []string `mapstructure:"ignore-check-signatures"`
+	ExtraStartSpanSignatures []string `mapstructure:"extra-start-span-signatures"`
 }
 
 type StaticCheckSettings struct {
-	// Deprecated: use the global `run.go` instead.
-	GoVersion string `mapstructure:"go"`
-
 	Checks                  []string `mapstructure:"checks"`
 	Initialisms             []string `mapstructure:"initialisms"`                // only for stylecheck
 	DotImportWhitelist      []string `mapstructure:"dot-import-whitelist"`       // only for stylecheck
 	HTTPStatusCodeWhitelist []string `mapstructure:"http-status-code-whitelist"` // only for stylecheck
+
+	// Deprecated: use the global `run.go` instead.
+	GoVersion string `mapstructure:"go"`
 }
 
 func (s *StaticCheckSettings) HasConfiguration() bool {
 	return len(s.Initialisms) > 0 || len(s.HTTPStatusCodeWhitelist) > 0 || len(s.DotImportWhitelist) > 0 || len(s.Checks) > 0
-}
-
-type StructCheckSettings struct {
-	CheckExportedFields bool `mapstructure:"exported-fields"`
 }
 
 type TagAlignSettings struct {
@@ -786,10 +913,31 @@ type TagAlignSettings struct {
 }
 
 type TagliatelleSettings struct {
-	Case struct {
-		Rules        map[string]string
-		UseFieldName bool `mapstructure:"use-field-name"`
-	}
+	Case TagliatelleCase
+}
+
+type TagliatelleCase struct {
+	TagliatelleBase `mapstructure:",squash"`
+	Overrides       []TagliatelleOverrides
+}
+
+type TagliatelleOverrides struct {
+	TagliatelleBase `mapstructure:",squash"`
+	Package         string `mapstructure:"pkg"`
+	Ignore          bool   `mapstructure:"ignore"`
+}
+
+type TagliatelleBase struct {
+	Rules         map[string]string                  `mapstructure:"rules"`
+	ExtendedRules map[string]TagliatelleExtendedRule `mapstructure:"extended-rules"`
+	UseFieldName  bool                               `mapstructure:"use-field-name"`
+	IgnoredFields []string                           `mapstructure:"ignored-fields"`
+}
+
+type TagliatelleExtendedRule struct {
+	Case                string
+	ExtraInitialisms    bool
+	InitialismOverrides map[string]bool
 }
 
 type TestifylintSettings struct {
@@ -798,9 +946,22 @@ type TestifylintSettings struct {
 	EnabledCheckers  []string `mapstructure:"enable"`
 	DisabledCheckers []string `mapstructure:"disable"`
 
+	BoolCompare struct {
+		IgnoreCustomTypes bool `mapstructure:"ignore-custom-types"`
+	} `mapstructure:"bool-compare"`
+
 	ExpectedActual struct {
 		ExpVarPattern string `mapstructure:"pattern"`
 	} `mapstructure:"expected-actual"`
+
+	Formatter struct {
+		CheckFormatString *bool `mapstructure:"check-format-string"`
+		RequireFFuncs     bool  `mapstructure:"require-f-funcs"`
+	} `mapstructure:"formatter"`
+
+	GoRequire struct {
+		IgnoreHTTPHandlers bool `mapstructure:"ignore-http-handlers"`
+	} `mapstructure:"go-require"`
 
 	RequireError struct {
 		FnPattern string `mapstructure:"fn-pattern"`
@@ -841,11 +1002,29 @@ type UseStdlibVarsSettings struct {
 	TimeLayout         bool `mapstructure:"time-layout"`
 	CryptoHash         bool `mapstructure:"crypto-hash"`
 	DefaultRPCPath     bool `mapstructure:"default-rpc-path"`
-	OSDevNull          bool `mapstructure:"os-dev-null"`
 	SQLIsolationLevel  bool `mapstructure:"sql-isolation-level"`
 	TLSSignatureScheme bool `mapstructure:"tls-signature-scheme"`
 	ConstantKind       bool `mapstructure:"constant-kind"`
-	SyslogPriority     bool `mapstructure:"syslog-priority"`
+
+	// Deprecated
+	OSDevNull *bool `mapstructure:"os-dev-null"`
+	// Deprecated
+	SyslogPriority *bool `mapstructure:"syslog-priority"`
+}
+
+type UseTestingSettings struct {
+	ContextBackground bool `mapstructure:"context-background"`
+	ContextTodo       bool `mapstructure:"context-todo"`
+	OSChdir           bool `mapstructure:"os-chdir"`
+	OSMkdirTemp       bool `mapstructure:"os-mkdir-temp"`
+	OSSetenv          bool `mapstructure:"os-setenv"`
+	OSTempDir         bool `mapstructure:"os-temp-dir"`
+	OSCreateTemp      bool `mapstructure:"os-create-temp"`
+}
+
+type UnconvertSettings struct {
+	FastMath bool `mapstructure:"fast-math"`
+	Safe     bool `mapstructure:"safe"`
 }
 
 type UnparamSettings struct {
@@ -856,15 +1035,13 @@ type UnparamSettings struct {
 type UnusedSettings struct {
 	FieldWritesAreUses     bool `mapstructure:"field-writes-are-uses"`
 	PostStatementsAreReads bool `mapstructure:"post-statements-are-reads"`
-	ExportedIsUsed         bool `mapstructure:"exported-is-used"`
 	ExportedFieldsAreUsed  bool `mapstructure:"exported-fields-are-used"`
 	ParametersAreUsed      bool `mapstructure:"parameters-are-used"`
 	LocalVariablesAreUsed  bool `mapstructure:"local-variables-are-used"`
 	GeneratedIsUsed        bool `mapstructure:"generated-is-used"`
-}
 
-type VarCheckSettings struct {
-	CheckExportedFields bool `mapstructure:"exported-fields"`
+	// Deprecated
+	ExportedIsUsed *bool `mapstructure:"exported-is-used"`
 }
 
 type VarnamelenSettings struct {
@@ -886,6 +1063,7 @@ type WhitespaceSettings struct {
 }
 
 type WrapcheckSettings struct {
+	ExtraIgnoreSigs []string `mapstructure:"extra-ignore-sigs"`
 	// TODO(ldez): v2 the options must be renamed to use hyphen.
 	IgnoreSigs             []string `mapstructure:"ignoreSigs"`
 	IgnoreSigRegexps       []string `mapstructure:"ignoreSigRegexps"`
@@ -910,17 +1088,15 @@ type WSLSettings struct {
 }
 
 // CustomLinterSettings encapsulates the meta-data of a private linter.
-// For example, a private linter may be added to the golangci config file as shown below.
-//
-//	linters-settings:
-//	 custom:
-//	   example:
-//	     path: /example.so
-//	     description: The description of the linter
-//	     original-url: github.com/golangci/example-linter
 type CustomLinterSettings struct {
+	// Type plugin type.
+	// It can be `goplugin` or `module`.
+	Type string `mapstructure:"type"`
+
 	// Path to a plugin *.so file that implements the private linter.
+	// Only for Go plugin system.
 	Path string
+
 	// Description describes the purpose of the private linter.
 	Description string
 	// OriginalURL The URL containing the source code for the private linter.
@@ -928,4 +1104,20 @@ type CustomLinterSettings struct {
 
 	// Settings plugin settings only work with linterdb.PluginConstructor symbol.
 	Settings any
+}
+
+func (s *CustomLinterSettings) Validate() error {
+	if s.Type == "module" {
+		if s.Path != "" {
+			return errors.New("path not supported with module type")
+		}
+
+		return nil
+	}
+
+	if s.Path == "" {
+		return errors.New("path is required")
+	}
+
+	return nil
 }
