@@ -12,10 +12,12 @@ func IsDir(filename string) bool {
 	return err == nil && fi.IsDir()
 }
 
-var cachedWd string
-var cachedWdError error
-var getWdOnce sync.Once
-var useCache = true
+var (
+	cachedWd      string
+	cachedWdError error
+	getWdOnce     sync.Once
+	useCache      = true
+)
 
 func UseWdCache(use bool) {
 	useCache = use
@@ -32,13 +34,13 @@ func Getwd() (string, error) {
 			return
 		}
 
-		evaledWd, err := EvalSymlinks(cachedWd)
+		evaluatedWd, err := EvalSymlinks(cachedWd)
 		if err != nil {
-			cachedWd, cachedWdError = "", fmt.Errorf("can't eval symlinks on wd %s: %s", cachedWd, err)
+			cachedWd, cachedWdError = "", fmt.Errorf("can't eval symlinks on wd %s: %w", cachedWd, err)
 			return
 		}
 
-		cachedWd = evaledWd
+		cachedWd = evaluatedWd
 	})
 
 	return cachedWd, cachedWdError
@@ -59,7 +61,7 @@ func EvalSymlinks(path string) (string, error) {
 	}
 
 	var er evalSymlinkRes
-	er.path, er.err = filepath.EvalSymlinks(path)
+	er.path, er.err = evalSymlinks(path)
 	evalSymlinkCache.Store(path, er)
 
 	return er.path, er.err
@@ -70,19 +72,19 @@ func ShortestRelPath(path, wd string) (string, error) {
 		var err error
 		wd, err = Getwd()
 		if err != nil {
-			return "", fmt.Errorf("can't get working directory: %s", err)
+			return "", fmt.Errorf("can't get working directory: %w", err)
 		}
 	}
 
-	evaledPath, err := EvalSymlinks(path)
+	evaluatedPath, err := EvalSymlinks(path)
 	if err != nil {
-		return "", fmt.Errorf("can't eval symlinks for path %s: %s", path, err)
+		return "", fmt.Errorf("can't eval symlinks for path %s: %w", path, err)
 	}
-	path = evaledPath
+	path = evaluatedPath
 
 	// make path absolute and then relative to be able to fix this case:
-	// we are in /test dir, we want to normalize ../test, and have file file.go in this dir;
-	// it must have normalized path file.go, not ../test/file.go,
+	// we are in `/test` dir, we want to normalize `../test`, and have file `file.go` in this dir;
+	// it must have normalized path `file.go`, not `../test/file.go`,
 	var absPath string
 	if filepath.IsAbs(path) {
 		absPath = path
@@ -92,7 +94,7 @@ func ShortestRelPath(path, wd string) (string, error) {
 
 	relPath, err := filepath.Rel(wd, absPath)
 	if err != nil {
-		return "", fmt.Errorf("can't get relative path for path %s and root %s: %s",
+		return "", fmt.Errorf("can't get relative path for path %s and root %s: %w",
 			absPath, wd, err)
 	}
 
