@@ -1,7 +1,8 @@
 package linter
 
 import (
-	"golang.org/x/tools/go/analysis"
+	"fmt"
+
 	"golang.org/x/tools/go/packages"
 
 	"github.com/golangci/golangci-lint/pkg/config"
@@ -26,10 +27,19 @@ const (
 // LastLinter nolintlint must be last because it looks at the results of all the previous linters for unused nolint directives.
 const LastLinter = "nolintlint"
 
+type DeprecationLevel int
+
+const (
+	DeprecationNone DeprecationLevel = iota
+	DeprecationWarning
+	DeprecationError
+)
+
 type Deprecation struct {
 	Since       string
 	Message     string
 	Replacement string
+	Level       DeprecationLevel
 }
 
 type Config struct {
@@ -71,7 +81,7 @@ func (lc *Config) IsSlowLinter() bool {
 }
 
 func (lc *Config) WithLoadFiles() *Config {
-	lc.LoadMode |= packages.NeedName | packages.NeedFiles | packages.NeedCompiledGoFiles
+	lc.LoadMode |= packages.NeedName | packages.NeedFiles | packages.NeedCompiledGoFiles | packages.NeedModule
 	return lc
 }
 
@@ -112,13 +122,22 @@ func (lc *Config) WithSince(version string) *Config {
 	return lc
 }
 
-func (lc *Config) Deprecated(message, version, replacement string) *Config {
+func (lc *Config) Deprecated(message, version, replacement string, level DeprecationLevel) *Config {
 	lc.Deprecation = &Deprecation{
 		Since:       version,
 		Message:     message,
 		Replacement: replacement,
+		Level:       level,
 	}
 	return lc
+}
+
+func (lc *Config) DeprecatedWarning(message, version, replacement string) *Config {
+	return lc.Deprecated(message, version, replacement, DeprecationWarning)
+}
+
+func (lc *Config) DeprecatedError(message, version, replacement string) *Config {
+	return lc.Deprecated(message, version, replacement, DeprecationError)
 }
 
 func (lc *Config) IsDeprecated() bool {
@@ -133,21 +152,29 @@ func (lc *Config) Name() string {
 	return lc.Linter.Name()
 }
 
-func (lc *Config) WithNoopFallback(cfg *config.Config) *Config {
-	if cfg != nil && config.IsGreaterThanOrEqualGo121(cfg.Run.Go) {
-		lc.Linter = &Noop{
-			name: lc.Linter.Name(),
-			desc: lc.Linter.Desc(),
-			run: func(pass *analysis.Pass) (any, error) {
-				return nil, nil
-			},
-		}
-
+func (lc *Config) WithNoopFallback(cfg *config.Config, cond func(cfg *config.Config) error) *Config {
+	if err := cond(cfg); err != nil {
+		lc.Linter = NewNoop(lc.Linter, err.Error())
 		lc.LoadMode = 0
+
 		return lc.WithLoadFiles()
 	}
 
 	return lc
+}
+
+func IsGoLowerThanGo122() func(cfg *config.Config) error {
+	return isGoLowerThanGo("1.22")
+}
+
+func isGoLowerThanGo(v string) func(cfg *config.Config) error {
+	return func(cfg *config.Config) error {
+		if cfg == nil || config.IsGoGreaterThanOrEqual(cfg.Run.Go, v) {
+			return nil
+		}
+
+		return fmt.Errorf("this linter is disabled because the Go version (%s) of your project is lower than Go %s", cfg.Run.Go, v)
+	}
 }
 
 func NewConfig(linter Linter) *Config {
