@@ -5,10 +5,9 @@ import (
 
 	"golang.org/x/tools/go/packages"
 
-	"github.com/golangci/golangci-lint/internal/pkgcache"
+	"github.com/golangci/golangci-lint/internal/cache"
 	"github.com/golangci/golangci-lint/pkg/config"
-	"github.com/golangci/golangci-lint/pkg/fsutils"
-	"github.com/golangci/golangci-lint/pkg/golinters/goanalysis/load"
+	"github.com/golangci/golangci-lint/pkg/goanalysis/load"
 	"github.com/golangci/golangci-lint/pkg/logutils"
 )
 
@@ -20,12 +19,10 @@ type Context struct {
 	// version for each of packages
 	OriginalPackages []*packages.Package
 
-	Cfg       *config.Config
-	FileCache *fsutils.FileCache
-	LineCache *fsutils.LineCache
-	Log       logutils.Log
+	Cfg *config.Config
+	Log logutils.Log
 
-	PkgCache  *pkgcache.Cache
+	PkgCache  *cache.Cache
 	LoadGuard *load.Guard
 }
 
