@@ -2,118 +2,163 @@ package lint
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"runtime/debug"
 	"strings"
 
-	"github.com/hashicorp/go-multierror"
-	gopackages "golang.org/x/tools/go/packages"
-
 	"github.com/golangci/golangci-lint/internal/errorutil"
 	"github.com/golangci/golangci-lint/pkg/config"
 	"github.com/golangci/golangci-lint/pkg/fsutils"
+	"github.com/golangci/golangci-lint/pkg/goformatters"
 	"github.com/golangci/golangci-lint/pkg/goutil"
 	"github.com/golangci/golangci-lint/pkg/lint/linter"
 	"github.com/golangci/golangci-lint/pkg/lint/lintersdb"
 	"github.com/golangci/golangci-lint/pkg/logutils"
-	"github.com/golangci/golangci-lint/pkg/packages"
 	"github.com/golangci/golangci-lint/pkg/result"
 	"github.com/golangci/golangci-lint/pkg/result/processors"
 	"github.com/golangci/golangci-lint/pkg/timeutils"
 )
 
-type Runner struct {
-	Processors []processors.Processor
-	Log        logutils.Log
+type processorStat struct {
+	inCount  int
+	outCount int
 }
 
-func NewRunner(cfg *config.Config, log logutils.Log, goenv *goutil.Env,
-	es *lintersdb.EnabledSet,
+type Runner struct {
+	Log logutils.Log
+
+	lintCtx    *linter.Context
+	Processors []processors.Processor
+}
+
+func NewRunner(log logutils.Log, cfg *config.Config, args []string, goenv *goutil.Env,
 	lineCache *fsutils.LineCache, fileCache *fsutils.FileCache,
-	dbManager *lintersdb.Manager, pkgs []*gopackages.Package) (*Runner, error) {
+	dbManager *lintersdb.Manager, lintCtx *linter.Context,
+) (*Runner, error) {
 	// Beware that some processors need to add the path prefix when working with paths
 	// because they get invoked before the path prefixer (exclude and severity rules)
 	// or process other paths (skip files).
 	files := fsutils.NewFiles(lineCache, cfg.Output.PathPrefix)
 
-	skipFilesProcessor, err := processors.NewSkipFiles(cfg.Run.SkipFiles, cfg.Output.PathPrefix)
+	pathRelativity, err := processors.NewPathRelativity(log, cfg.GetBasePath())
+	if err != nil {
+		return nil, fmt.Errorf("error creating path relativity processor: %w", err)
+	}
+
+	exclusionPaths, err := processors.NewExclusionPaths(log, &cfg.Linters.LinterExclusions)
 	if err != nil {
 		return nil, err
 	}
 
-	skipDirs := cfg.Run.SkipDirs
-	if cfg.Run.UseDefaultSkipDirs {
-		skipDirs = append(skipDirs, packages.StdExcludeDirRegexps...)
-	}
-	skipDirsProcessor, err := processors.NewSkipDirs(skipDirs, log.Child(logutils.DebugKeySkipDirs), cfg.Run.Args, cfg.Output.PathPrefix)
+	skipFilesProcessor, err := processors.NewSkipFiles(cfg.Issues.ExcludeFiles, cfg.Output.PathPrefix)
 	if err != nil {
 		return nil, err
 	}
 
-	enabledLinters, err := es.GetEnabledLintersMap()
+	skipDirs := cfg.Issues.ExcludeDirs
+	if cfg.Issues.UseDefaultExcludeDirs {
+		skipDirs = append(skipDirs, processors.StdExcludeDirRegexps...)
+	}
+
+	skipDirsProcessor, err := processors.NewSkipDirs(log.Child(logutils.DebugKeySkipDirs), skipDirs, args, cfg.Output.PathPrefix)
+	if err != nil {
+		return nil, err
+	}
+
+	enabledLinters, err := dbManager.GetEnabledLintersMap()
 	if err != nil {
 		return nil, fmt.Errorf("failed to get enabled linters: %w", err)
 	}
 
-	// print deprecated messages
-	if !cfg.InternalCmdTest {
-		for name, lc := range enabledLinters {
-			if !lc.IsDeprecated() {
-				continue
-			}
-
-			var extra string
-			if lc.Deprecation.Replacement != "" {
-				extra = fmt.Sprintf("Replaced by %s.", lc.Deprecation.Replacement)
-			}
-
-			log.Warnf("The linter '%s' is deprecated (since %s) due to: %s %s", name, lc.Deprecation.Since, lc.Deprecation.Message, extra)
-		}
+	metaFormatter, err := goformatters.NewMetaFormatter(log, cfg, enabledLinters)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create meta-formatter: %w", err)
 	}
 
 	return &Runner{
 		Processors: []processors.Processor{
+			// Must be the first processor.
+			processors.NewPathAbsoluter(log),
+
 			processors.NewCgo(goenv),
 
-			// Must go after Cgo.
-			processors.NewFilenameUnadjuster(pkgs, log.Child(logutils.DebugKeyFilenameUnadjuster)),
+			// Must be after Cgo.
+			processors.NewFilenameUnadjuster(lintCtx.Packages, log.Child(logutils.DebugKeyFilenameUnadjuster)),
 
-			// Must be before diff, nolint and exclude autogenerated processor at least.
-			processors.NewPathPrettifier(),
+			// Must be after FilenameUnadjuster.
+			processors.NewInvalidIssue(log.Child(logutils.DebugKeyInvalidIssue)),
+
+			// Must be after PathAbsoluter, Cgo, FilenameUnadjuster InvalidIssue.
+			pathRelativity,
+
+			// Must be after PathRelativity.
+			exclusionPaths,
 			skipFilesProcessor,
-			skipDirsProcessor, // must be after path prettifier
+			skipDirsProcessor,
 
-			processors.NewAutogeneratedExclude(),
+			processors.NewGeneratedFileFilter(cfg.Linters.LinterExclusions.Generated),
 
 			// Must be before exclude because users see already marked output and configure excluding by it.
 			processors.NewIdentifierMarker(),
 
-			getExcludeProcessor(&cfg.Issues),
-			getExcludeRulesProcessor(&cfg.Issues, log, files),
-			processors.NewNolint(log.Child(logutils.DebugKeyNolint), dbManager, enabledLinters),
+			processors.NewExclusionRules(log.Child(logutils.DebugKeyExclusionRules), files,
+				&cfg.Linters.LinterExclusions, &cfg.Issues),
 
-			processors.NewUniqByLine(cfg),
-			processors.NewDiff(cfg.Issues.Diff, cfg.Issues.DiffFromRevision, cfg.Issues.DiffPatchFilePath, cfg.Issues.WholeFiles),
+			processors.NewNolintFilter(log.Child(logutils.DebugKeyNolintFilter), dbManager, enabledLinters),
+
+			processors.NewDiff(&cfg.Issues),
+
+			// The fixer still needs to see paths for the issues that are relative to the current directory.
+			processors.NewFixer(cfg, log, fileCache, metaFormatter),
+
+			// Must be after the Fixer.
+			processors.NewUniqByLine(cfg.Issues.UniqByLine),
 			processors.NewMaxPerFileFromLinter(cfg),
 			processors.NewMaxSameIssues(cfg.Issues.MaxSameIssues, log.Child(logutils.DebugKeyMaxSameIssues), cfg),
 			processors.NewMaxFromLinter(cfg.Issues.MaxIssuesPerLinter, log.Child(logutils.DebugKeyMaxFromLinter), cfg),
-			processors.NewSourceCode(lineCache, log.Child(logutils.DebugKeySourceCode)),
-			processors.NewPathShortener(),
-			getSeverityRulesProcessor(&cfg.Severity, log, files),
-
-			// The fixer still needs to see paths for the issues that are relative to the current directory.
-			processors.NewFixer(cfg, log, fileCache),
 
 			// Now we can modify the issues for output.
-			processors.NewPathPrefixer(cfg.Output.PathPrefix),
-			processors.NewSortResults(cfg),
+			processors.NewSourceCode(lineCache, log.Child(logutils.DebugKeySourceCode)),
+			processors.NewPathShortener(),
+			processors.NewSeverity(log.Child(logutils.DebugKeySeverityRules), files, &cfg.Severity),
+			processors.NewPathPrettifier(log, cfg.Output.PathPrefix),
+			processors.NewSortResults(&cfg.Output),
 		},
-		Log: log,
+		lintCtx: lintCtx,
+		Log:     log,
 	}, nil
 }
 
+func (r *Runner) Run(ctx context.Context, linters []*linter.Config) ([]result.Issue, error) {
+	sw := timeutils.NewStopwatch("linters", r.Log)
+	defer sw.Print()
+
+	var (
+		lintErrors error
+		issues     []result.Issue
+	)
+
+	for _, lc := range linters {
+		linterIssues, err := timeutils.TrackStage(sw, lc.Name(), func() ([]result.Issue, error) {
+			return r.runLinterSafe(ctx, r.lintCtx, lc)
+		})
+		if err != nil {
+			lintErrors = errors.Join(lintErrors, fmt.Errorf("can't run linter %s", lc.Linter.Name()), err)
+			r.Log.Warnf("Can't run linter %s: %v", lc.Linter.Name(), err)
+
+			continue
+		}
+
+		issues = append(issues, linterIssues...)
+	}
+
+	return r.processLintResults(issues), lintErrors
+}
+
 func (r *Runner) runLinterSafe(ctx context.Context, lintCtx *linter.Context,
-	lc *linter.Config) (ret []result.Issue, err error) {
+	lc *linter.Config,
+) (ret []result.Issue, err error) {
 	defer func() {
 		if panicData := recover(); panicData != nil {
 			if pe, ok := panicData.(*errorutil.PanicError); ok {
@@ -152,12 +197,7 @@ func (r *Runner) runLinterSafe(ctx context.Context, lintCtx *linter.Context,
 	return issues, nil
 }
 
-type processorStat struct {
-	inCount  int
-	outCount int
-}
-
-func (r Runner) processLintResults(inIssues []result.Issue) []result.Issue {
+func (r *Runner) processLintResults(inIssues []result.Issue) []result.Issue {
 	sw := timeutils.NewStopwatch("processing", r.Log)
 
 	var issuesBefore, issuesAfter int
@@ -173,10 +213,7 @@ func (r Runner) processLintResults(inIssues []result.Issue) []result.Issue {
 	// finalize processors: logging, clearing, no heavy work here
 
 	for _, p := range r.Processors {
-		p := p
-		sw.TrackStage(p.Name(), func() {
-			p.Finish()
-		})
+		sw.TrackStage(p.Name(), p.Finish)
 	}
 
 	if issuesBefore != issuesAfter {
@@ -188,52 +225,22 @@ func (r Runner) processLintResults(inIssues []result.Issue) []result.Issue {
 	return outIssues
 }
 
-func (r Runner) printPerProcessorStat(stat map[string]processorStat) {
+func (r *Runner) printPerProcessorStat(stat map[string]processorStat) {
 	parts := make([]string, 0, len(stat))
 	for name, ps := range stat {
 		if ps.inCount != 0 {
-			parts = append(parts, fmt.Sprintf("%s: %d/%d", name, ps.outCount, ps.inCount))
+			parts = append(parts, fmt.Sprintf("%s: %d/%d", name, ps.inCount, ps.outCount))
 		}
 	}
 	if len(parts) != 0 {
-		r.Log.Infof("Processors filtering stat (out/in): %s", strings.Join(parts, ", "))
+		r.Log.Infof("Processors filtering stat (in/out): %s", strings.Join(parts, ", "))
 	}
-}
-
-func (r Runner) Run(ctx context.Context, linters []*linter.Config, lintCtx *linter.Context) ([]result.Issue, error) {
-	sw := timeutils.NewStopwatch("linters", r.Log)
-	defer sw.Print()
-
-	var (
-		lintErrors *multierror.Error
-		issues     []result.Issue
-	)
-
-	for _, lc := range linters {
-		lc := lc
-		sw.TrackStage(lc.Name(), func() {
-			linterIssues, err := r.runLinterSafe(ctx, lintCtx, lc)
-			if err != nil {
-				lintErrors = multierror.Append(lintErrors, fmt.Errorf("can't run linter %s: %w", lc.Linter.Name(), err))
-				r.Log.Warnf("Can't run linter %s: %v", lc.Linter.Name(), err)
-
-				return
-			}
-
-			issues = append(issues, linterIssues...)
-		})
-	}
-
-	return r.processLintResults(issues), lintErrors.ErrorOrNil()
 }
 
 func (r *Runner) processIssues(issues []result.Issue, sw *timeutils.Stopwatch, statPerProcessor map[string]processorStat) []result.Issue {
 	for _, p := range r.Processors {
-		var newIssues []result.Issue
-		var err error
-		p := p
-		sw.TrackStage(p.Name(), func() {
-			newIssues, err = p.Process(issues)
+		newIssues, err := timeutils.TrackStage(sw, p.Name(), func() ([]result.Issue, error) {
+			return p.Process(issues)
 		})
 
 		if err != nil {
@@ -246,105 +253,11 @@ func (r *Runner) processIssues(issues []result.Issue, sw *timeutils.Stopwatch, s
 			issues = newIssues
 		}
 
+		// This is required by JSON serialization
 		if issues == nil {
 			issues = []result.Issue{}
 		}
 	}
 
 	return issues
-}
-
-func getExcludeProcessor(cfg *config.Issues) processors.Processor {
-	var excludeTotalPattern string
-
-	if len(cfg.ExcludePatterns) != 0 {
-		excludeTotalPattern = fmt.Sprintf("(%s)", strings.Join(cfg.ExcludePatterns, "|"))
-	}
-
-	var excludeProcessor processors.Processor
-	if cfg.ExcludeCaseSensitive {
-		excludeProcessor = processors.NewExcludeCaseSensitive(excludeTotalPattern)
-	} else {
-		excludeProcessor = processors.NewExclude(excludeTotalPattern)
-	}
-
-	return excludeProcessor
-}
-
-func getExcludeRulesProcessor(cfg *config.Issues, log logutils.Log, files *fsutils.Files) processors.Processor {
-	var excludeRules []processors.ExcludeRule
-	for _, r := range cfg.ExcludeRules {
-		excludeRules = append(excludeRules, processors.ExcludeRule{
-			BaseRule: processors.BaseRule{
-				Text:       r.Text,
-				Source:     r.Source,
-				Path:       r.Path,
-				PathExcept: r.PathExcept,
-				Linters:    r.Linters,
-			},
-		})
-	}
-
-	if cfg.UseDefaultExcludes {
-		for _, r := range config.GetExcludePatterns(cfg.IncludeDefaultExcludes) {
-			excludeRules = append(excludeRules, processors.ExcludeRule{
-				BaseRule: processors.BaseRule{
-					Text:    r.Pattern,
-					Linters: []string{r.Linter},
-				},
-			})
-		}
-	}
-
-	var excludeRulesProcessor processors.Processor
-	if cfg.ExcludeCaseSensitive {
-		excludeRulesProcessor = processors.NewExcludeRulesCaseSensitive(
-			excludeRules,
-			files,
-			log.Child(logutils.DebugKeyExcludeRules),
-		)
-	} else {
-		excludeRulesProcessor = processors.NewExcludeRules(
-			excludeRules,
-			files,
-			log.Child(logutils.DebugKeyExcludeRules),
-		)
-	}
-
-	return excludeRulesProcessor
-}
-
-func getSeverityRulesProcessor(cfg *config.Severity, log logutils.Log, files *fsutils.Files) processors.Processor {
-	var severityRules []processors.SeverityRule
-	for _, r := range cfg.Rules {
-		severityRules = append(severityRules, processors.SeverityRule{
-			Severity: r.Severity,
-			BaseRule: processors.BaseRule{
-				Text:       r.Text,
-				Source:     r.Source,
-				Path:       r.Path,
-				PathExcept: r.PathExcept,
-				Linters:    r.Linters,
-			},
-		})
-	}
-
-	var severityRulesProcessor processors.Processor
-	if cfg.CaseSensitive {
-		severityRulesProcessor = processors.NewSeverityRulesCaseSensitive(
-			cfg.Default,
-			severityRules,
-			files,
-			log.Child(logutils.DebugKeySeverityRules),
-		)
-	} else {
-		severityRulesProcessor = processors.NewSeverityRules(
-			cfg.Default,
-			severityRules,
-			files,
-			log.Child(logutils.DebugKeySeverityRules),
-		)
-	}
-
-	return severityRulesProcessor
 }
