@@ -5,6 +5,9 @@ import (
 	"strings"
 )
 
+// EnvTestRun value: "1"
+const EnvTestRun = "GL_TEST_RUN"
+
 // envDebug value: one or several debug keys.
 // examples:
 // - Remove output to `/dev/null`: `GL_DEBUG=linters_output ./golangci-lint run`
@@ -13,35 +16,53 @@ import (
 const envDebug = "GL_DEBUG"
 
 const (
-	DebugKeyAutogenExclude     = "autogen_exclude" // Debugs a filter excluding autogenerated source code.
-	DebugKeyBinSalt            = "bin_salt"
-	DebugKeyConfigReader       = "config_reader"
-	DebugKeyEmpty              = ""
-	DebugKeyEnabledLinters     = "enabled_linters"
-	DebugKeyEnv                = "env" // Debugs `go env` command.
-	DebugKeyExcludeRules       = "exclude_rules"
-	DebugKeyExec               = "exec"
-	DebugKeyFilenameUnadjuster = "filename_unadjuster"
-	DebugKeyForbidigo          = "forbidigo"
-	DebugKeyGoEnv              = "goenv"
-	DebugKeyLinter             = "linter"
-	DebugKeyLintersContext     = "linters_context"
-	DebugKeyLintersDB          = "lintersdb"
-	DebugKeyLintersOutput      = "linters_output"
-	DebugKeyLoader             = "loader" // Debugs packages loading (including `go/packages` internal debugging).
-	DebugKeyMaxFromLinter      = "max_from_linter"
-	DebugKeyMaxSameIssues      = "max_same_issues"
-	DebugKeyPkgCache           = "pkgcache"
-	DebugKeyRunner             = "runner"
-	DebugKeySeverityRules      = "severity_rules"
-	DebugKeySkipDirs           = "skip_dirs"
-	DebugKeySourceCode         = "source_code"
-	DebugKeyStopwatch          = "stopwatch"
+	DebugKeyBinSalt        = "bin_salt"
+	DebugKeyConfigReader   = "config_reader"
+	DebugKeyEmpty          = ""
+	DebugKeyEnabledLinters = "enabled_linters"
+	DebugKeyExec           = "exec"
+	DebugKeyFormatter      = "formatter"
+	DebugKeyGoEnv          = "goenv"
+	DebugKeyLinter         = "linter"
+	DebugKeyLintersContext = "linters_context"
+	DebugKeyLintersDB      = "lintersdb"
+	DebugKeyLintersOutput  = "linters_output"
+	DebugKeyLoader         = "loader" // Debugs packages loading (including `go/packages` internal debugging).
+	DebugKeyPkgCache       = "pkgcache"
+	DebugKeyRunner         = "runner"
+	DebugKeyStopwatch      = "stopwatch"
+	DebugKeyTest           = "test"
+)
+
+// Printers.
+const (
+	DebugKeyCheckstylePrinter  = "checkstyle_printer"
+	DebugKeyCodeClimatePrinter = "codeclimate_printer"
+	DebugKeySarifPrinter       = "sarif_printer"
 	DebugKeyTabPrinter         = "tab_printer"
-	DebugKeyTest               = "test"
+	DebugKeyTeamCityPrinter    = "teamcity_printer"
 	DebugKeyTextPrinter        = "text_printer"
 )
 
+// Processors.
+const (
+	DebugKeyExclusionPaths      = "exclusion_paths"
+	DebugKeyExclusionRules      = "exclusion_rules"
+	DebugKeyFilenameUnadjuster  = "filename_unadjuster"
+	DebugKeyGeneratedFileFilter = "generated_file_filter" // Debugs a filter excluding autogenerated source code.
+	DebugKeyInvalidIssue        = "invalid_issue"
+	DebugKeyMaxFromLinter       = "max_from_linter"
+	DebugKeyMaxSameIssues       = "max_same_issues"
+	DebugKeyNolintFilter        = "nolint_filter" // Debugs a filter excluding issues by `//nolint` comments.
+	DebugKeyPathAbsoluter       = "path_absoluter"
+	DebugKeyPathPrettifier      = "path_prettifier"
+	DebugKeyPathRelativity      = "path_relativity"
+	DebugKeySeverityRules       = "severity_rules"
+	DebugKeySkipDirs            = "skip_dirs"
+	DebugKeySourceCode          = "source_code"
+)
+
+// Analysis.
 const (
 	DebugKeyGoAnalysis = "goanalysis"
 
@@ -55,10 +76,11 @@ const (
 	DebugKeyGoAnalysisFactsInherit = DebugKeyGoAnalysisFacts + "/inherit"
 )
 
+// Linters.
 const (
-	DebugKeyGoCritic  = "gocritic"  // Debugs `go-critic` linter.
-	DebugKeyMegacheck = "megacheck" // Debugs `staticcheck` related linters.
-	DebugKeyNolint    = "nolint"    // Debugs a filter excluding issues by `//nolint` comments.
+	DebugKeyForbidigo = "forbidigo" // Debugs `forbidigo` linter.
+	DebugKeyGoCritic  = "gocritic"  // Debugs `gocritic` linter.
+	DebugKeyGovet     = "govet"     // Debugs `govet` linter.
 	DebugKeyRevive    = "revive"    // Debugs `revive` linter.
 )
 
@@ -99,8 +121,15 @@ func HaveDebugTag(tag string) bool {
 	return enabledDebugs[tag]
 }
 
+var verbose bool
+
 func SetupVerboseLog(log Log, isVerbose bool) {
 	if isVerbose {
+		verbose = isVerbose
 		log.SetLevel(LogLevelInfo)
 	}
+}
+
+func IsVerbose() bool {
+	return verbose
 }
