@@ -9,6 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/golangci/golangci-lint/pkg/logutils"
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
@@ -41,16 +42,67 @@ func TestCheckstyle_Print(t *testing.T) {
 				Column:   9,
 			},
 		},
+		{
+			FromLinter: "linter-c",
+			Severity:   "",
+			Text:       "without severity",
+			SourceLines: []string{
+				"func foo() {",
+				"\tfmt.Println(\"bar\")",
+				"}",
+			},
+			Pos: token.Position{
+				Filename: "path/to/filec.go",
+				Offset:   5,
+				Line:     300,
+				Column:   10,
+			},
+		},
+		{
+			FromLinter: "linter-d",
+			Severity:   "foo",
+			Text:       "unknown severity",
+			SourceLines: []string{
+				"func foo() {",
+				"\tfmt.Println(\"bar\")",
+				"}",
+			},
+			Pos: token.Position{
+				Filename: "path/to/filed.go",
+				Offset:   5,
+				Line:     300,
+				Column:   11,
+			},
+		},
 	}
 
 	buf := new(bytes.Buffer)
-	printer := NewCheckstyle(buf)
+
+	log := logutils.NewStderrLog(logutils.DebugKeyEmpty)
+	log.SetLevel(logutils.LogLevelDebug)
+
+	printer := NewCheckstyle(log, buf)
 
 	err := printer.Print(issues)
 	require.NoError(t, err)
 
-	//nolint:lll
-	expected := "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n\n<checkstyle version=\"5.0\">\n  <file name=\"path/to/filea.go\">\n    <error column=\"4\" line=\"10\" message=\"some issue\" severity=\"warning\" source=\"linter-a\"></error>\n  </file>\n  <file name=\"path/to/fileb.go\">\n    <error column=\"9\" line=\"300\" message=\"another issue\" severity=\"error\" source=\"linter-b\"></error>\n  </file>\n</checkstyle>\n"
+	expected := `<?xml version="1.0" encoding="UTF-8"?>
+
+<checkstyle version="5.0">
+  <file name="path/to/filea.go">
+    <error column="4" line="10" message="some issue" severity="warning" source="linter-a"></error>
+  </file>
+  <file name="path/to/fileb.go">
+    <error column="9" line="300" message="another issue" severity="error" source="linter-b"></error>
+  </file>
+  <file name="path/to/filec.go">
+    <error column="10" line="300" message="without severity" severity="error" source="linter-c"></error>
+  </file>
+  <file name="path/to/filed.go">
+    <error column="11" line="300" message="unknown severity" severity="error" source="linter-d"></error>
+  </file>
+</checkstyle>
+`
 
 	assert.Equal(t, expected, strings.ReplaceAll(buf.String(), "\r", ""))
 }
