@@ -11,6 +11,7 @@ import (
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
+// Tab prints issues using tabulation as a field separator.
 type Tab struct {
 	printLinterName bool
 	useColors       bool
@@ -19,11 +20,11 @@ type Tab struct {
 	w   io.Writer
 }
 
-func NewTab(printLinterName, useColors bool, log logutils.Log, w io.Writer) *Tab {
+func NewTab(log logutils.Log, w io.Writer, printLinterName, useColors bool) *Tab {
 	return &Tab{
 		printLinterName: printLinterName,
 		useColors:       useColors,
-		log:             log,
+		log:             log.Child(logutils.DebugKeyTabPrinter),
 		w:               w,
 	}
 }
@@ -52,15 +53,15 @@ func (p *Tab) Print(issues []result.Issue) error {
 	return nil
 }
 
-func (p *Tab) printIssue(i *result.Issue, w io.Writer) {
-	text := p.SprintfColored(color.FgRed, "%s", i.Text)
+func (p *Tab) printIssue(issue *result.Issue, w io.Writer) {
+	text := p.SprintfColored(color.FgRed, "%s", issue.Text)
 	if p.printLinterName {
-		text = fmt.Sprintf("%s\t%s", i.FromLinter, text)
+		text = fmt.Sprintf("%s\t%s", issue.FromLinter, text)
 	}
 
-	pos := p.SprintfColored(color.Bold, "%s:%d", i.FilePath(), i.Line())
-	if i.Pos.Column != 0 {
-		pos += fmt.Sprintf(":%d", i.Pos.Column)
+	pos := p.SprintfColored(color.Bold, "%s:%d", issue.FilePath(), issue.Line())
+	if issue.Pos.Column != 0 {
+		pos += fmt.Sprintf(":%d", issue.Pos.Column)
 	}
 
 	fmt.Fprintf(w, "%s\t%s\n", pos, text)
