@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"io"
 	"strings"
-	"unicode/utf8"
 
+	"github.com/golangci/golangci-lint/pkg/logutils"
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
@@ -15,16 +15,22 @@ const (
 	largeLimit = 4000
 )
 
-// TeamCity printer for TeamCity format.
+const defaultTeamCitySeverity = "ERROR"
+
+// TeamCity prints issues in the TeamCity format.
+// https://www.jetbrains.com/help/teamcity/service-messages.html
 type TeamCity struct {
-	w       io.Writer
-	escaper *strings.Replacer
+	log       logutils.Log
+	w         io.Writer
+	escaper   *strings.Replacer
+	sanitizer severitySanitizer
 }
 
 // NewTeamCity output format outputs issues according to TeamCity service message format.
-func NewTeamCity(w io.Writer) *TeamCity {
+func NewTeamCity(log logutils.Log, w io.Writer) *TeamCity {
 	return &TeamCity{
-		w: w,
+		log: log.Child(logutils.DebugKeyTeamCityPrinter),
+		w:   w,
 		// https://www.jetbrains.com/help/teamcity/service-messages.html#Escaped+Values
 		escaper: strings.NewReplacer(
 			"'", "|'",
@@ -34,6 +40,11 @@ func NewTeamCity(w io.Writer) *TeamCity {
 			"[", "|[",
 			"]", "|]",
 		),
+		sanitizer: severitySanitizer{
+			// https://www.jetbrains.com/help/teamcity/service-messages.html#Inspection+Instance
+			allowedSeverities: []string{"INFO", defaultTeamCitySeverity, "WARNING", "WEAK WARNING"},
+			defaultSeverity:   defaultTeamCitySeverity,
+		},
 	}
 }
 
@@ -65,13 +76,18 @@ func (p *TeamCity) Print(issues []result.Issue) error {
 			message:  issue.Text,
 			file:     issue.FilePath(),
 			line:     issue.Line(),
-			severity: issue.Severity,
+			severity: p.sanitizer.Sanitize(strings.ToUpper(issue.Severity)),
 		}
 
 		_, err := instance.Print(p.w, p.escaper)
 		if err != nil {
 			return err
 		}
+	}
+
+	err := p.sanitizer.Err()
+	if err != nil {
+		p.log.Infof("%v", err)
 	}
 
 	return nil
@@ -88,8 +104,8 @@ type InspectionType struct {
 }
 
 func (i InspectionType) Print(w io.Writer, escaper *strings.Replacer) (int, error) {
-	return fmt.Fprintf(w, "##teamcity[InspectionType id='%s' name='%s' description='%s' category='%s']\n",
-		limit(i.id, smallLimit), limit(i.name, smallLimit), limit(escaper.Replace(i.description), largeLimit), limit(i.category, smallLimit))
+	return fmt.Fprintf(w, "##teamcity[inspectionType id='%s' name='%s' description='%s' category='%s']\n",
+		cutVal(i.id, smallLimit), cutVal(i.name, smallLimit), cutVal(escaper.Replace(i.description), largeLimit), cutVal(i.category, smallLimit))
 }
 
 // InspectionInstance reports a specific defect, warning, error message.
@@ -105,18 +121,16 @@ type InspectionInstance struct {
 
 func (i InspectionInstance) Print(w io.Writer, replacer *strings.Replacer) (int, error) {
 	return fmt.Fprintf(w, "##teamcity[inspection typeId='%s' message='%s' file='%s' line='%d' SEVERITY='%s']\n",
-		limit(i.typeID, smallLimit),
-		limit(replacer.Replace(i.message), largeLimit),
-		limit(i.file, largeLimit),
-		i.line, strings.ToUpper(i.severity))
+		cutVal(i.typeID, smallLimit),
+		cutVal(replacer.Replace(i.message), largeLimit),
+		cutVal(i.file, largeLimit),
+		i.line, i.severity)
 }
 
-func limit(s string, max int) string {
-	var size, count int
-	for i := 0; i < max && count < len(s); i++ {
-		_, size = utf8.DecodeRuneInString(s[count:])
-		count += size
+func cutVal(s string, limit int) string {
+	runes := []rune(s)
+	if len(runes) > limit {
+		return string(runes[:limit])
 	}
-
-	return s[:count]
+	return s
 }
