@@ -11,21 +11,22 @@ import (
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
+// Text prints issues with a human friendly representation.
 type Text struct {
-	printIssuedLine bool
 	printLinterName bool
+	printIssuedLine bool
 	useColors       bool
 
 	log logutils.Log
 	w   io.Writer
 }
 
-func NewText(printIssuedLine, useColors, printLinterName bool, log logutils.Log, w io.Writer) *Text {
+func NewText(log logutils.Log, w io.Writer, printLinterName, printIssuedLine, useColors bool) *Text {
 	return &Text{
-		printIssuedLine: printIssuedLine,
 		printLinterName: printLinterName,
+		printIssuedLine: printIssuedLine,
 		useColors:       useColors,
-		log:             log,
+		log:             log.Child(logutils.DebugKeyTextPrinter),
 		w:               w,
 	}
 }
@@ -55,32 +56,32 @@ func (p *Text) Print(issues []result.Issue) error {
 	return nil
 }
 
-func (p *Text) printIssue(i *result.Issue) {
-	text := p.SprintfColored(color.FgRed, "%s", strings.TrimSpace(i.Text))
+func (p *Text) printIssue(issue *result.Issue) {
+	text := p.SprintfColored(color.FgRed, "%s", strings.TrimSpace(issue.Text))
 	if p.printLinterName {
-		text += fmt.Sprintf(" (%s)", i.FromLinter)
+		text += fmt.Sprintf(" (%s)", issue.FromLinter)
 	}
-	pos := p.SprintfColored(color.Bold, "%s:%d", i.FilePath(), i.Line())
-	if i.Pos.Column != 0 {
-		pos += fmt.Sprintf(":%d", i.Pos.Column)
+	pos := p.SprintfColored(color.Bold, "%s:%d", issue.FilePath(), issue.Line())
+	if issue.Pos.Column != 0 {
+		pos += fmt.Sprintf(":%d", issue.Pos.Column)
 	}
 	fmt.Fprintf(p.w, "%s: %s\n", pos, text)
 }
 
-func (p *Text) printSourceCode(i *result.Issue) {
-	for _, line := range i.SourceLines {
+func (p *Text) printSourceCode(issue *result.Issue) {
+	for _, line := range issue.SourceLines {
 		fmt.Fprintln(p.w, line)
 	}
 }
 
-func (p *Text) printUnderLinePointer(i *result.Issue) {
+func (p *Text) printUnderLinePointer(issue *result.Issue) {
 	// if column == 0 it means column is unknown (e.g. for gosec)
-	if len(i.SourceLines) != 1 || i.Pos.Column == 0 {
+	if len(issue.SourceLines) != 1 || issue.Pos.Column == 0 {
 		return
 	}
 
-	col0 := i.Pos.Column - 1
-	line := i.SourceLines[0]
+	col0 := issue.Pos.Column - 1
+	line := issue.SourceLines[0]
 	prefixRunes := make([]rune, 0, len(line))
 	for j := 0; j < len(line) && j < col0; j++ {
 		if line[j] == '\t' {
