@@ -3,18 +3,13 @@ package processors
 import (
 	"regexp"
 
+	"github.com/golangci/golangci-lint/pkg/config"
 	"github.com/golangci/golangci-lint/pkg/fsutils"
 	"github.com/golangci/golangci-lint/pkg/logutils"
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
-type BaseRule struct {
-	Text       string
-	Source     string
-	Path       string
-	PathExcept string
-	Linters    []string
-}
+const caseInsensitivePrefix = "(?i)"
 
 type baseRule struct {
 	text       *regexp.Regexp
@@ -22,6 +17,32 @@ type baseRule struct {
 	path       *regexp.Regexp
 	pathExcept *regexp.Regexp
 	linters    []string
+}
+
+// The usage of `regexp.MustCompile()` is safe here,
+// because the regular expressions are checked before inside [config.BaseRule.Validate].
+func newBaseRule(rule *config.BaseRule, prefix string) baseRule {
+	base := baseRule{
+		linters: rule.Linters,
+	}
+
+	if rule.Text != "" {
+		base.text = regexp.MustCompile(prefix + rule.Text)
+	}
+
+	if rule.Source != "" {
+		base.source = regexp.MustCompile(prefix + rule.Source)
+	}
+
+	if rule.Path != "" {
+		base.path = regexp.MustCompile(fsutils.NormalizePathInRegex(rule.Path))
+	}
+
+	if rule.PathExcept != "" {
+		base.pathExcept = regexp.MustCompile(fsutils.NormalizePathInRegex(rule.PathExcept))
+	}
+
+	return base
 }
 
 func (r *baseRule) isEmpty() bool {
@@ -35,10 +56,10 @@ func (r *baseRule) match(issue *result.Issue, files *fsutils.Files, log logutils
 	if r.text != nil && !r.text.MatchString(issue.Text) {
 		return false
 	}
-	if r.path != nil && !r.path.MatchString(files.WithPathPrefix(issue.FilePath())) {
+	if r.path != nil && !r.path.MatchString(files.WithPathPrefix(issue.RelativePath)) {
 		return false
 	}
-	if r.pathExcept != nil && r.pathExcept.MatchString(issue.FilePath()) {
+	if r.pathExcept != nil && r.pathExcept.MatchString(issue.RelativePath) {
 		return false
 	}
 	if len(r.linters) != 0 && !r.matchLinter(issue) {
@@ -63,12 +84,26 @@ func (r *baseRule) matchLinter(issue *result.Issue) bool {
 	return false
 }
 
-func (r *baseRule) matchSource(issue *result.Issue, lineCache *fsutils.LineCache, log logutils.Log) bool { //nolint:interfacer
-	sourceLine, errSourceLine := lineCache.GetLine(issue.FilePath(), issue.Line())
+func (r *baseRule) matchSource(issue *result.Issue, lineCache *fsutils.LineCache, log logutils.Log) bool {
+	sourceLine, errSourceLine := lineCache.GetLine(issue.RelativePath, issue.Line())
 	if errSourceLine != nil {
-		log.Warnf("Failed to get line %s:%d from line cache: %s", issue.FilePath(), issue.Line(), errSourceLine)
+		log.Warnf("Failed to get line %s:%d from line cache: %s", issue.RelativePath, issue.Line(), errSourceLine)
 		return false // can't properly match
 	}
 
 	return r.source.MatchString(sourceLine)
+}
+
+func parseRules[T, V any](rules []T, prefix string, newFn func(*T, string) V) []V {
+	if len(rules) == 0 {
+		return nil
+	}
+
+	parsedRules := make([]V, 0, len(rules))
+
+	for _, r := range rules {
+		parsedRules = append(parsedRules, newFn(&r, prefix))
+	}
+
+	return parsedRules
 }
