@@ -5,15 +5,13 @@ import (
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
-type linterToCountMap map[string]int
-type fileToLinterToCountMap map[string]linterToCountMap
+var _ Processor = (*MaxPerFileFromLinter)(nil)
 
+// MaxPerFileFromLinter limits the number of reports by file and by linter.
 type MaxPerFileFromLinter struct {
-	flc                        fileToLinterToCountMap
+	fileLinterCounter          fileLinterCounter
 	maxPerFileFromLinterConfig map[string]int
 }
-
-var _ Processor = &MaxPerFileFromLinter{}
 
 func NewMaxPerFileFromLinter(cfg *config.Config) *MaxPerFileFromLinter {
 	maxPerFileFromLinterConfig := map[string]int{}
@@ -23,37 +21,55 @@ func NewMaxPerFileFromLinter(cfg *config.Config) *MaxPerFileFromLinter {
 		// otherwise we need to fix all issues in the file at once
 		maxPerFileFromLinterConfig["gofmt"] = 1
 		maxPerFileFromLinterConfig["goimports"] = 1
+		maxPerFileFromLinterConfig["gci"] = 1
 	}
 
 	return &MaxPerFileFromLinter{
-		flc:                        fileToLinterToCountMap{},
+		fileLinterCounter:          fileLinterCounter{},
 		maxPerFileFromLinterConfig: maxPerFileFromLinterConfig,
 	}
 }
 
-func (p *MaxPerFileFromLinter) Name() string {
+func (*MaxPerFileFromLinter) Name() string {
 	return "max_per_file_from_linter"
 }
 
 func (p *MaxPerFileFromLinter) Process(issues []result.Issue) ([]result.Issue, error) {
-	return filterIssues(issues, func(i *result.Issue) bool {
-		limit := p.maxPerFileFromLinterConfig[i.FromLinter]
+	return filterIssuesUnsafe(issues, func(issue *result.Issue) bool {
+		limit := p.maxPerFileFromLinterConfig[issue.FromLinter]
 		if limit == 0 {
 			return true
 		}
 
-		lm := p.flc[i.FilePath()]
-		if lm == nil {
-			p.flc[i.FilePath()] = linterToCountMap{}
-		}
-		count := p.flc[i.FilePath()][i.FromLinter]
-		if count >= limit {
+		if p.fileLinterCounter.GetCount(issue) >= limit {
 			return false
 		}
 
-		p.flc[i.FilePath()][i.FromLinter]++
+		p.fileLinterCounter.Increment(issue)
+
 		return true
 	}), nil
 }
 
-func (p *MaxPerFileFromLinter) Finish() {}
+func (*MaxPerFileFromLinter) Finish() {}
+
+type fileLinterCounter map[string]map[string]int
+
+func (f fileLinterCounter) GetCount(issue *result.Issue) int {
+	return f.getCounter(issue)[issue.FromLinter]
+}
+
+func (f fileLinterCounter) Increment(issue *result.Issue) {
+	f.getCounter(issue)[issue.FromLinter]++
+}
+
+func (f fileLinterCounter) getCounter(issue *result.Issue) map[string]int {
+	lc := f[issue.FilePath()]
+
+	if lc == nil {
+		lc = map[string]int{}
+		f[issue.FilePath()] = lc
+	}
+
+	return lc
+}
