@@ -8,27 +8,26 @@ import (
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
-type textToCountMap map[string]int
+var _ Processor = (*MaxSameIssues)(nil)
 
+// MaxSameIssues limits the number of reports with the same text.
 type MaxSameIssues struct {
-	tc    textToCountMap
-	limit int
-	log   logutils.Log
-	cfg   *config.Config
+	textCounter map[string]int
+	limit       int
+	log         logutils.Log
+	cfg         *config.Config
 }
-
-var _ Processor = &MaxSameIssues{}
 
 func NewMaxSameIssues(limit int, log logutils.Log, cfg *config.Config) *MaxSameIssues {
 	return &MaxSameIssues{
-		tc:    textToCountMap{},
-		limit: limit,
-		log:   log,
-		cfg:   cfg,
+		textCounter: map[string]int{},
+		limit:       limit,
+		log:         log,
+		cfg:         cfg,
 	}
 }
 
-func (p *MaxSameIssues) Name() string {
+func (*MaxSameIssues) Name() string {
 	return "max_same_issues"
 }
 
@@ -37,19 +36,15 @@ func (p *MaxSameIssues) Process(issues []result.Issue) ([]result.Issue, error) {
 		return issues, nil
 	}
 
-	return filterIssues(issues, func(i *result.Issue) bool {
-		if i.Replacement != nil && p.cfg.Issues.NeedFix {
-			// we need to fix all issues at once => we need to return all of them
-			return true
-		}
+	return filterIssuesUnsafe(issues, func(issue *result.Issue) bool {
+		p.textCounter[issue.Text]++ // always inc for stat
 
-		p.tc[i.Text]++ // always inc for stat
-		return p.tc[i.Text] <= p.limit
+		return p.textCounter[issue.Text] <= p.limit
 	}), nil
 }
 
 func (p *MaxSameIssues) Finish() {
-	walkStringToIntMapSortedByValue(p.tc, func(text string, count int) {
+	walkStringToIntMapSortedByValue(p.textCounter, func(text string, count int) {
 		if count > p.limit {
 			p.log.Infof("%d/%d issues with text %q were hidden, use --max-same-issues",
 				count-p.limit, count, text)
