@@ -8,33 +8,34 @@ import (
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
+var _ Processor = (*PathShortener)(nil)
+
+// PathShortener modifies text of the reports to reduce file path inside the text.
+// It uses the rooted path name corresponding to the current directory (`wd`).
 type PathShortener struct {
 	wd string
 }
-
-var _ Processor = PathShortener{}
 
 func NewPathShortener() *PathShortener {
 	wd, err := fsutils.Getwd()
 	if err != nil {
 		panic(fmt.Sprintf("Can't get working dir: %s", err))
 	}
-	return &PathShortener{
-		wd: wd,
-	}
+
+	return &PathShortener{wd: wd}
 }
 
-func (p PathShortener) Name() string {
+func (PathShortener) Name() string {
 	return "path_shortener"
 }
 
 func (p PathShortener) Process(issues []result.Issue) ([]result.Issue, error) {
-	return transformIssues(issues, func(i *result.Issue) *result.Issue {
-		newI := i
-		newI.Text = strings.ReplaceAll(newI.Text, p.wd+"/", "")
-		newI.Text = strings.ReplaceAll(newI.Text, p.wd, "")
-		return newI
+	return transformIssues(issues, func(issue *result.Issue) *result.Issue {
+		newIssue := issue
+		newIssue.Text = strings.ReplaceAll(newIssue.Text, p.wd+"/", "")
+		newIssue.Text = strings.ReplaceAll(newIssue.Text, p.wd, "")
+		return newIssue
 	}), nil
 }
 
-func (p PathShortener) Finish() {}
+func (PathShortener) Finish() {}
