@@ -4,70 +4,71 @@ import (
 	"fmt"
 	"path/filepath"
 	"regexp"
-	"strings"
 
 	"github.com/golangci/golangci-lint/pkg/fsutils"
 	"github.com/golangci/golangci-lint/pkg/logutils"
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
+var _ Processor = (*SkipDirs)(nil)
+
+var StdExcludeDirRegexps = []string{
+	normalizePathRegex("vendor"),
+	normalizePathRegex("third_party"),
+	normalizePathRegex("testdata"),
+	normalizePathRegex("examples"),
+	normalizePathRegex("Godeps"),
+	normalizePathRegex("builtin"),
+}
+
 type skipStat struct {
 	pattern string
 	count   int
 }
 
+// SkipDirs filters reports based on directory names.
+// It uses the shortest relative paths and `path-prefix` option.
+// TODO(ldez): should be removed in v2.
 type SkipDirs struct {
-	patterns         []*regexp.Regexp
-	log              logutils.Log
+	patterns   []*regexp.Regexp
+	pathPrefix string
+
+	log logutils.Log
+
 	skippedDirs      map[string]*skipStat
 	absArgsDirs      []string
 	skippedDirsCache map[string]bool
-	pathPrefix       string
 }
 
-var _ Processor = (*SkipDirs)(nil)
-
-const goFileSuffix = ".go"
-
-func NewSkipDirs(patterns []string, log logutils.Log, runArgs []string, pathPrefix string) (*SkipDirs, error) {
+func NewSkipDirs(log logutils.Log, patterns, args []string, pathPrefix string) (*SkipDirs, error) {
 	var patternsRe []*regexp.Regexp
 	for _, p := range patterns {
 		p = fsutils.NormalizePathInRegex(p)
+
 		patternRe, err := regexp.Compile(p)
 		if err != nil {
 			return nil, fmt.Errorf("can't compile regexp %q: %w", p, err)
 		}
+
 		patternsRe = append(patternsRe, patternRe)
 	}
 
-	if len(runArgs) == 0 {
-		runArgs = append(runArgs, "./...")
-	}
-	var absArgsDirs []string
-	for _, arg := range runArgs {
-		base := filepath.Base(arg)
-		if base == "..." || strings.HasSuffix(base, goFileSuffix) {
-			arg = filepath.Dir(arg)
-		}
-
-		absArg, err := filepath.Abs(arg)
-		if err != nil {
-			return nil, fmt.Errorf("failed to abs-ify arg %q: %w", arg, err)
-		}
-		absArgsDirs = append(absArgsDirs, absArg)
+	absArgsDirs, err := absDirs(args)
+	if err != nil {
+		return nil, err
 	}
 
 	return &SkipDirs{
 		patterns:         patternsRe,
+		pathPrefix:       pathPrefix,
 		log:              log,
 		skippedDirs:      map[string]*skipStat{},
 		absArgsDirs:      absArgsDirs,
 		skippedDirsCache: map[string]bool{},
-		pathPrefix:       pathPrefix,
 	}, nil
 }
 
-func (p *SkipDirs) Name() string {
+func (*SkipDirs) Name() string {
 	return "skip_dirs"
 }
 
@@ -79,31 +80,34 @@ func (p *SkipDirs) Process(issues []result.Issue) ([]result.Issue, error) {
 	return filterIssues(issues, p.shouldPassIssue), nil
 }
 
-func (p *SkipDirs) shouldPassIssue(i *result.Issue) bool {
-	if filepath.IsAbs(i.FilePath()) {
-		if !isSpecialAutogeneratedFile(i.FilePath()) {
-			p.log.Warnf("Got abs path %s in skip dirs processor, it should be relative", i.FilePath())
-		}
-		return true
+func (p *SkipDirs) Finish() {
+	for dir, stat := range p.skippedDirs {
+		p.log.Infof("Skipped %d issues from dir %s by pattern %s", stat.count, dir, stat.pattern)
 	}
+}
 
-	issueRelDir := filepath.Dir(i.FilePath())
+func (p *SkipDirs) shouldPassIssue(issue *result.Issue) bool {
+	issueRelDir := filepath.Dir(issue.RelativePath)
 
 	if toPass, ok := p.skippedDirsCache[issueRelDir]; ok {
 		if !toPass {
 			p.skippedDirs[issueRelDir].count++
 		}
+
 		return toPass
 	}
 
 	issueAbsDir, err := filepath.Abs(issueRelDir)
 	if err != nil {
 		p.log.Warnf("Can't abs-ify path %q: %s", issueRelDir, err)
+
 		return true
 	}
 
 	toPass := p.shouldPassIssueDirs(issueRelDir, issueAbsDir)
+
 	p.skippedDirsCache[issueRelDir] = toPass
+
 	return toPass
 }
 
@@ -123,15 +127,19 @@ func (p *SkipDirs) shouldPassIssueDirs(issueRelDir, issueAbsDir string) bool {
 	// disadvantages (https://github.com/golangci/golangci-lint/pull/313).
 
 	path := fsutils.WithPathPrefix(p.pathPrefix, issueRelDir)
+
 	for _, pattern := range p.patterns {
 		if pattern.MatchString(path) {
 			ps := pattern.String()
+
 			if p.skippedDirs[issueRelDir] == nil {
 				p.skippedDirs[issueRelDir] = &skipStat{
 					pattern: ps,
 				}
 			}
+
 			p.skippedDirs[issueRelDir].count++
+
 			return false
 		}
 	}
@@ -139,8 +147,34 @@ func (p *SkipDirs) shouldPassIssueDirs(issueRelDir, issueAbsDir string) bool {
 	return true
 }
 
-func (p *SkipDirs) Finish() {
-	for dir, stat := range p.skippedDirs {
-		p.log.Infof("Skipped %d issues from dir %s by pattern %s", stat.count, dir, stat.pattern)
+func absDirs(args []string) ([]string, error) {
+	if len(args) == 0 {
+		args = append(args, "./...")
 	}
+
+	var absArgsDirs []string
+	for _, arg := range args {
+		base := filepath.Base(arg)
+		if base == "..." || isGoFile(base) {
+			arg = filepath.Dir(arg)
+		}
+
+		absArg, err := filepath.Abs(arg)
+		if err != nil {
+			return nil, fmt.Errorf("failed to abs-ify arg %q: %w", arg, err)
+		}
+
+		absArgsDirs = append(absArgsDirs, absArg)
+	}
+
+	return absArgsDirs, nil
+}
+
+func normalizePathRegex(e string) string {
+	return createPathRegex(e, filepath.Separator)
+}
+
+func createPathRegex(e string, sep rune) string {
+	escapedSep := regexp.QuoteMeta(string(sep)) // needed for windows sep '\\'
+	return fmt.Sprintf(`(^|%[1]s)%[2]s($|%[1]s)`, escapedSep, e)
 }
