@@ -1,166 +1,147 @@
 package processors
 
 import (
-	"sort"
+	"cmp"
+	"fmt"
+	"slices"
 	"strings"
 
 	"github.com/golangci/golangci-lint/pkg/config"
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
-// Base propose of this functionality to sort results (issues)
-// produced by various linters by analyzing code. We're achieving this
-// by sorting results.Issues using processor step, and chain based
-// rules that can compare different properties of the Issues struct.
+const (
+	orderNameFile     = "file"
+	orderNameLinter   = "linter"
+	orderNameSeverity = "severity"
+)
+
+const (
+	less = iota - 1
+	equal
+	greater
+)
 
 var _ Processor = (*SortResults)(nil)
 
+type issueComparator func(a, b *result.Issue) int
+
+// SortResults sorts reports based on criteria:
+//   - file names, line numbers, positions
+//   - linter names
+//   - severity names
 type SortResults struct {
-	cmp comparator
-	cfg *config.Config
+	cmps map[string][]issueComparator
+
+	cfg *config.Output
 }
 
-func NewSortResults(cfg *config.Config) *SortResults {
-	// For sorting we are comparing (in next order): file names, line numbers,
-	// position, and finally - giving up.
+func NewSortResults(cfg *config.Output) *SortResults {
 	return &SortResults{
-		cmp: ByName{
-			next: ByLine{
-				next: ByColumn{},
-			},
+		cmps: map[string][]issueComparator{
+			// For sorting we are comparing (in next order):
+			// file names, line numbers, position, and finally - giving up.
+			orderNameFile: {byFileName, byLine, byColumn},
+			// For sorting we are comparing: linter name
+			orderNameLinter: {byLinter},
+			// For sorting we are comparing: severity
+			orderNameSeverity: {bySeverity},
 		},
 		cfg: cfg,
 	}
 }
 
+func (SortResults) Name() string { return "sort_results" }
+
 // Process is performing sorting of the result issues.
-func (sr SortResults) Process(issues []result.Issue) ([]result.Issue, error) {
-	if !sr.cfg.Output.SortResults {
+func (p SortResults) Process(issues []result.Issue) ([]result.Issue, error) {
+	if !p.cfg.SortResults {
 		return issues, nil
 	}
 
-	sort.Slice(issues, func(i, j int) bool {
-		return sr.cmp.Compare(&issues[i], &issues[j]) == Less
+	if len(p.cfg.SortOrder) == 0 {
+		p.cfg.SortOrder = []string{orderNameFile}
+	}
+
+	var cmps []issueComparator
+
+	for _, name := range p.cfg.SortOrder {
+		c, ok := p.cmps[name]
+		if !ok {
+			return nil, fmt.Errorf("unsupported sort-order name %q", name)
+		}
+
+		cmps = append(cmps, c...)
+	}
+
+	comp := mergeComparators(cmps...)
+
+	slices.SortFunc(issues, func(a, b result.Issue) int {
+		return comp(&a, &b)
 	})
 
 	return issues, nil
 }
 
-func (sr SortResults) Name() string { return "sort_results" }
-func (sr SortResults) Finish()      {}
+func (SortResults) Finish() {}
 
-type compareResult int
-
-const (
-	Less compareResult = iota - 1
-	Equal
-	Greater
-	None
-)
-
-func (c compareResult) isNeutral() bool {
-	// return true if compare result is incomparable or equal.
-	return c == None || c == Equal
+func byFileName(a, b *result.Issue) int {
+	return strings.Compare(a.FilePath(), b.FilePath())
 }
 
-func (c compareResult) String() string {
-	switch c {
-	case Less:
-		return "Less"
-	case Equal:
-		return "Equal"
-	case Greater:
-		return "Greater"
-	}
-
-	return "None"
+func byLine(a, b *result.Issue) int {
+	return numericCompare(a.Line(), b.Line())
 }
 
-// comparator describe how to implement compare for two "issues" lexicographically
-type comparator interface {
-	Compare(a, b *result.Issue) compareResult
-	Next() comparator
+func byColumn(a, b *result.Issue) int {
+	return numericCompare(a.Column(), b.Column())
 }
 
-var (
-	_ comparator = (*ByName)(nil)
-	_ comparator = (*ByLine)(nil)
-	_ comparator = (*ByColumn)(nil)
-)
-
-type ByName struct{ next comparator }
-
-func (cmp ByName) Next() comparator { return cmp.next }
-
-func (cmp ByName) Compare(a, b *result.Issue) compareResult {
-	var res compareResult
-
-	if res = compareResult(strings.Compare(a.FilePath(), b.FilePath())); !res.isNeutral() {
-		return res
-	}
-
-	if next := cmp.Next(); next != nil {
-		return next.Compare(a, b)
-	}
-
-	return res
+func byLinter(a, b *result.Issue) int {
+	return strings.Compare(a.FromLinter, b.FromLinter)
 }
 
-type ByLine struct{ next comparator }
-
-func (cmp ByLine) Next() comparator { return cmp.next }
-
-func (cmp ByLine) Compare(a, b *result.Issue) compareResult {
-	var res compareResult
-
-	if res = numericCompare(a.Line(), b.Line()); !res.isNeutral() {
-		return res
-	}
-
-	if next := cmp.Next(); next != nil {
-		return next.Compare(a, b)
-	}
-
-	return res
+func bySeverity(a, b *result.Issue) int {
+	return severityCompare(a.Severity, b.Severity)
 }
 
-type ByColumn struct{ next comparator }
+func severityCompare(a, b string) int {
+	// The position inside the slice define the importance (lower to higher).
+	classic := []string{"low", "medium", "high", "warning", "error"}
 
-func (cmp ByColumn) Next() comparator { return cmp.next }
-
-func (cmp ByColumn) Compare(a, b *result.Issue) compareResult {
-	var res compareResult
-
-	if res = numericCompare(a.Column(), b.Column()); !res.isNeutral() {
-		return res
+	if slices.Contains(classic, a) && slices.Contains(classic, b) {
+		return cmp.Compare(slices.Index(classic, a), slices.Index(classic, b))
 	}
 
-	if next := cmp.Next(); next != nil {
-		return next.Compare(a, b)
+	if slices.Contains(classic, a) {
+		return greater
 	}
 
-	return res
+	if slices.Contains(classic, b) {
+		return less
+	}
+
+	return strings.Compare(a, b)
 }
 
-func numericCompare(a, b int) compareResult {
-	var (
-		isValuesInvalid  = a < 0 || b < 0
-		isZeroValuesBoth = a == 0 && b == 0
-		isEqual          = a == b
-		isZeroValueInA   = b > 0 && a == 0
-		isZeroValueInB   = a > 0 && b == 0
-	)
-
-	switch {
-	case isZeroValuesBoth || isEqual:
-		return Equal
-	case isValuesInvalid || isZeroValueInA || isZeroValueInB:
-		return None
-	case a > b:
-		return Greater
-	case a < b:
-		return Less
+func numericCompare(a, b int) int {
+	// Negative values and zeros are skipped (equal) because they either invalid or  "neutral" (default int value).
+	if a <= 0 || b <= 0 {
+		return equal
 	}
 
-	return Equal
+	return cmp.Compare(a, b)
+}
+
+func mergeComparators(comps ...issueComparator) issueComparator {
+	return func(a, b *result.Issue) int {
+		for _, comp := range comps {
+			i := comp(a, b)
+			if i != equal {
+				return i
+			}
+		}
+
+		return equal
+	}
 }
