@@ -6,12 +6,19 @@ import (
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
+var _ Processor = (*SourceCode)(nil)
+
+// SourceCode modifies displayed information based on [result.Issue.GetLineRange()].
+//
+// This is used:
+//   - to display the "UnderLinePointer".
+//   - in some rare cases to display multiple lines instead of one (ex: `dupl`)
+//
+// It requires to use [fsutils.LineCache] ([fsutils.FileCache]) to get the file information before the fixes.
 type SourceCode struct {
 	lineCache *fsutils.LineCache
 	log       logutils.Log
 }
-
-var _ Processor = SourceCode{}
 
 func NewSourceCode(lc *fsutils.LineCache, log logutils.Log) *SourceCode {
 	return &SourceCode{
@@ -20,28 +27,31 @@ func NewSourceCode(lc *fsutils.LineCache, log logutils.Log) *SourceCode {
 	}
 }
 
-func (p SourceCode) Name() string {
+func (SourceCode) Name() string {
 	return "source_code"
 }
 
 func (p SourceCode) Process(issues []result.Issue) ([]result.Issue, error) {
-	return transformIssues(issues, func(i *result.Issue) *result.Issue {
-		newI := *i
-
-		lineRange := i.GetLineRange()
-		for lineNumber := lineRange.From; lineNumber <= lineRange.To; lineNumber++ {
-			line, err := p.lineCache.GetLine(i.FilePath(), lineNumber)
-			if err != nil {
-				p.log.Warnf("Failed to get line %d for file %s: %s",
-					lineNumber, i.FilePath(), err)
-				return i
-			}
-
-			newI.SourceLines = append(newI.SourceLines, line)
-		}
-
-		return &newI
-	}), nil
+	return transformIssues(issues, p.transform), nil
 }
 
-func (p SourceCode) Finish() {}
+func (SourceCode) Finish() {}
+
+func (p SourceCode) transform(issue *result.Issue) *result.Issue {
+	newIssue := *issue
+
+	lineRange := issue.GetLineRange()
+	for lineNumber := lineRange.From; lineNumber <= lineRange.To; lineNumber++ {
+		line, err := p.lineCache.GetLine(issue.FilePath(), lineNumber)
+		if err != nil {
+			p.log.Warnf("Failed to get line %d for file %s: %s",
+				lineNumber, issue.FilePath(), err)
+
+			return issue
+		}
+
+		newIssue.SourceLines = append(newIssue.SourceLines, line)
+	}
+
+	return &newIssue
+}
