@@ -1,58 +1,67 @@
 package processors
 
 import (
-	"github.com/golangci/golangci-lint/pkg/config"
 	"github.com/golangci/golangci-lint/pkg/result"
 )
 
-type lineToCount map[int]int
-type fileToLineToCount map[string]lineToCount
+const uniqByLineLimit = 1
 
+var _ Processor = (*UniqByLine)(nil)
+
+// UniqByLine filters reports to keep only one report by line of code.
 type UniqByLine struct {
-	flc fileToLineToCount
-	cfg *config.Config
+	fileLineCounter fileLineCounter
+	enabled         bool
 }
 
-func NewUniqByLine(cfg *config.Config) *UniqByLine {
+func NewUniqByLine(enable bool) *UniqByLine {
 	return &UniqByLine{
-		flc: fileToLineToCount{},
-		cfg: cfg,
+		fileLineCounter: fileLineCounter{},
+		enabled:         enable,
 	}
 }
 
-var _ Processor = &UniqByLine{}
-
-func (p *UniqByLine) Name() string {
+func (*UniqByLine) Name() string {
 	return "uniq_by_line"
 }
 
 func (p *UniqByLine) Process(issues []result.Issue) ([]result.Issue, error) {
-	if !p.cfg.Output.UniqByLine {
+	if !p.enabled {
 		return issues, nil
 	}
 
-	return filterIssues(issues, func(i *result.Issue) bool {
-		if i.Replacement != nil && p.cfg.Issues.NeedFix {
-			// if issue will be auto-fixed we shouldn't collapse issues:
-			// e.g. one line can contain 2 misspellings, they will be in 2 issues and misspell should fix both of them.
-			return true
-		}
-
-		lc := p.flc[i.FilePath()]
-		if lc == nil {
-			lc = lineToCount{}
-			p.flc[i.FilePath()] = lc
-		}
-
-		const limit = 1
-		count := lc[i.Line()]
-		if count == limit {
-			return false
-		}
-
-		lc[i.Line()]++
-		return true
-	}), nil
+	return filterIssuesUnsafe(issues, p.shouldPassIssue), nil
 }
 
-func (p *UniqByLine) Finish() {}
+func (*UniqByLine) Finish() {}
+
+func (p *UniqByLine) shouldPassIssue(issue *result.Issue) bool {
+	if p.fileLineCounter.GetCount(issue) == uniqByLineLimit {
+		return false
+	}
+
+	p.fileLineCounter.Increment(issue)
+
+	return true
+}
+
+type fileLineCounter map[string]map[int]int
+
+func (f fileLineCounter) GetCount(issue *result.Issue) int {
+	return f.getCounter(issue)[issue.Line()]
+}
+
+func (f fileLineCounter) Increment(issue *result.Issue) {
+	f.getCounter(issue)[issue.Line()]++
+}
+
+func (f fileLineCounter) getCounter(issue *result.Issue) map[int]int {
+	lc := f[issue.FilePath()]
+
+	if lc == nil {
+		lc = map[int]int{}
+		f[issue.FilePath()] = lc
+	}
+
+	return lc
+}
