@@ -12,11 +12,11 @@ import (
 	"github.com/golangci/golangci-lint/test/testshared"
 )
 
-//nolint:misspell,lll
-const expectedJSONOutput = `{"Issues":[{"FromLinter":"misspell","Text":"` + "`" + `occured` + "`" + ` is a misspelling of ` + "`" + `occurred` + "`" + `","Severity":"","SourceLines":["\t// comment with incorrect spelling: occured // want \"` + "`" + `occured` + "`" + ` is a misspelling of ` + "`" + `occurred` + "`" + `\""],"Replacement":{"NeedOnlyDelete":false,"NewLines":null,"Inline":{"StartCol":37,"Length":7,"NewString":"occurred"}},"Pos":{"Filename":"testdata/misspell.go","Offset":0,"Line":6,"Column":38},"ExpectNoLint":false,"ExpectedNoLintLinter":""}]`
+//nolint:misspell // misspelling is intentional
+const expectedJSONOutput = `{"Issues":[{"FromLinter":"misspell","Text":"` + "`" + `occured` + "`" + ` is a misspelling of ` + "`" + `occurred` + "`" + `","Severity":"","SourceLines":["\t// comment with incorrect spelling: occured // want \"` + "`" + `occured` + "`" + ` is a misspelling of ` + "`" + `occurred` + "`" + `\""],"Pos":{"Filename":"testdata/output.go","Offset":159,"Line":6,"Column":38},"SuggestedFixes":[{"Message":"","TextEdits":[{"Pos":159,"End":166,"NewText":"b2NjdXJyZWQ="}]}],"ExpectNoLint":false,"ExpectedNoLintLinter":""}]`
 
 func TestOutput_lineNumber(t *testing.T) {
-	sourcePath := filepath.Join(testdataDir, "misspell.go")
+	sourcePath := filepath.Join(testdataDir, "output.go")
 
 	testshared.NewRunnerBuilder(t).
 		WithArgs(
@@ -30,12 +30,12 @@ func TestOutput_lineNumber(t *testing.T) {
 		Runner().
 		Install().
 		Run().
-		//nolint:misspell
-		ExpectHasIssue("testdata/misspell.go:6:38: `occured` is a misspelling of `occurred`")
+		//nolint:misspell // misspelling is intentional
+		ExpectHasIssue("testdata/output.go:6:38: `occured` is a misspelling of `occurred`")
 }
 
 func TestOutput_Stderr(t *testing.T) {
-	sourcePath := filepath.Join(testdataDir, "misspell.go")
+	sourcePath := filepath.Join(testdataDir, "output.go")
 
 	testshared.NewRunnerBuilder(t).
 		WithArgs(
@@ -55,7 +55,7 @@ func TestOutput_Stderr(t *testing.T) {
 func TestOutput_File(t *testing.T) {
 	resultPath := filepath.Join(t.TempDir(), "golangci_lint_test_result")
 
-	sourcePath := filepath.Join(testdataDir, "misspell.go")
+	sourcePath := filepath.Join(testdataDir, "output.go")
 
 	testshared.NewRunnerBuilder(t).
 		WithArgs(
@@ -77,7 +77,7 @@ func TestOutput_File(t *testing.T) {
 }
 
 func TestOutput_Multiple(t *testing.T) {
-	sourcePath := filepath.Join(testdataDir, "misspell.go")
+	sourcePath := filepath.Join(testdataDir, "output.go")
 
 	testshared.NewRunnerBuilder(t).
 		WithArgs(
@@ -91,7 +91,7 @@ func TestOutput_Multiple(t *testing.T) {
 		Runner().
 		Install().
 		Run().
-		//nolint:misspell
-		ExpectHasIssue("testdata/misspell.go:6:38: `occured` is a misspelling of `occurred`").
+		//nolint:misspell // misspelling is intentional
+		ExpectHasIssue("testdata/output.go:6:38: `occured` is a misspelling of `occurred`").
 		ExpectOutputContains(testshared.NormalizeFilePathInJSON(expectedJSONOutput))
 }
