@@ -9,6 +9,8 @@ import (
 	"testing"
 )
 
+const binaryName = "golangci-lint.exe"
+
 // SkipOnWindows skip test on Windows.
 func SkipOnWindows(tb testing.TB) {
 	tb.Skip("not supported on Windows")
@@ -26,11 +28,6 @@ func NormalizeFilePathInJSON(in string) string {
 // NormalizeFileInString normalizes in quoted string, ie. replace `\\` with `\\\\`.
 func NormalizeFileInString(in string) string {
 	return strings.ReplaceAll(filepath.FromSlash(in), "\\", "\\\\")
-}
-
-// defaultBinaryName returns the path to the default binary.
-func defaultBinaryName() string {
-	return filepath.Join("..", "golangci-lint.exe")
 }
 
 // normalizeFilePath find Go file path and replace `/` with `\\`.
